+++ src/Main.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
 
-public class Main extends data
+public class Main extends com_moves
 {
 
     MouseListener ml = new MouseListener()
@@ -762,213 +762,7 @@ public class Main extends data
 
 
 
-    public void hud_change()
-    {
-        if(play_ai_count > 0 && play_ai_count < 5)
-        {
-            p_air.setIcon(found);
-            if(play_ai_count != air_temp_c)
-            {
-                air_temp_c = play_ai_count;
-                explosion();
-            }
-        }
-        if(play_ai_count == 5)
-        {
-            p_air.setIcon(hit);
-            if(play_ai_count != air_temp_c)
-            {
-                air_temp_c = play_ai_count;
-                explosion();
-            }
-        }
-        if(play_bat_count > 0 && play_bat_count < 4)
-        {
-            p_bat.setIcon(found);
-            if(play_bat_count != bat_temp_c)
-            {
-                bat_temp_c = play_bat_count;
-                explosion();
-            }
-        }
-        if(play_bat_count == 4)
-        {
-            p_bat.setIcon(hit);
-            if(play_bat_count != bat_temp_c)
-            {
-                bat_temp_c = play_bat_count;
-                explosion();
-            }
-        }
-        if(play_sub_count > 0 && play_sub_count < 3)
-        {
-            p_sub.setIcon(found);
-            if(play_sub_count != sub_temp_c)
-            {
-                sub_temp_c = play_sub_count;
-                explosion();
-            }
-        }
-        if(play_sub_count == 3)
-        {
-            p_sub.setIcon(hit);
-            if(play_sub_count != sub_temp_c)
-            {
-                sub_temp_c = play_sub_count;
-                explosion();
-            }
-        }
-        if(play_cru_count > 0 && play_cru_count < 3)
-        {
-            p_cru.setIcon(found);
-            if(play_cru_count != cru_temp_c)
-            {
-                cru_temp_c = play_cru_count;
-                explosion();
-            }
-        }
-        if(play_cru_count == 3)
-        {
-            p_cru.setIcon(hit);
-            if(play_cru_count != cru_temp_c)
-            {
-                cru_temp_c = play_cru_count;
-                explosion();
-            }
-        }
-        if(play_des_count > 0 && play_des_count < 2)
-        {
-            p_des.setIcon(found);
-            if(play_des_count != des_temp_c)
-            {
-                des_temp_c = play_des_count;
-                explosion();
-            }
-        }
-        if(play_des_count == 2)
-        {
-            p_des.setIcon(hit);
-            if(play_des_count != des_temp_c)
-            {
-                des_temp_c = play_des_count;
-                explosion();
-            }
-        }
-        if(com_ai_count > 0 && com_ai_count < 5)
-        {
-            c_air.setIcon(found);
-            if(com_ai_count != air_temp_p)
-            {
-                air_temp_p = com_ai_count;
-                explosion();
-            }
-        }
-        if(com_ai_count == 5)
-        {
-            c_air.setIcon(safe);
-            if(com_ai_count != air_temp_p)
-            {
-                air_temp_p = com_ai_count;
-                explosion();
-            }
-        }
-        if(com_bat_count > 0 && com_bat_count < 4)
-        {
-            c_bat.setIcon(found);
-            if(com_bat_count != bat_temp_p)
-            {
-                bat_temp_p = com_bat_count;
-                explosion();
-            }
-        }
-        if(com_bat_count == 4)
-        {
-            c_bat.setIcon(safe);
-            if(com_bat_count != bat_temp_p)
-            {
-                bat_temp_p = com_bat_count;
-                explosion();
-            }
-        }
-        if(com_sub_count > 0 && com_sub_count < 3)
-        {
-            c_sub.setIcon(found);
-            if(com_sub_count != sub_temp_p)
-            {
-                sub_temp_p = com_sub_count;
-                explosion();
-            }
-        }
-        if(com_sub_count == 3)
-        {
-            c_sub.setIcon(safe);
-            if(com_sub_count != sub_temp_p)
-            {
-                sub_temp_p = com_sub_count;
-                explosion();
-            }
-        }
-        if(com_cru_count > 0 && com_cru_count < 3)
-        {
-            c_cru.setIcon(found);
-            if(com_cru_count != cru_temp_p)
-            {
-                cru_temp_p = com_cru_count;
-                explosion();
-            }
-        }
-        if(com_cru_count == 3)
-        {
-            c_cru.setIcon(safe);
-            if(com_cru_count != cru_temp_p)
-            {
-                cru_temp_p = com_cru_count;
-                explosion();
-            }
-        }
-        if(com_des_count > 0 && com_des_count < 2)
-        {
-            c_des.setIcon(found);
-            if(com_des_count != des_temp_p)
-            {
-                des_temp_p = com_des_count;
-                explosion();
-            }
-        }
-        if(com_des_count == 2)
-        {
-            c_des.setIcon(safe);
-            if(com_des_count != des_temp_p)
-            {
-                des_temp_p = com_des_count;
-                explosion();
-            }
-        }
-    }
-    public void explosion()
-    {
-        try
-        {
-            ac.setFramePosition(0);
-            ac.start();
-        }
-        catch (Exception ex)
-        {
-            ex.printStackTrace();
-        }
-    }
-    public void miss()
-    {
-        try
-        {
-            acm.setFramePosition(0);
-            acm.start();
-        }
-        catch (Exception ex)
-        {
-            ex.printStackTrace();
-        }
-    }
+
     public void l1Call()
     {
         int flag = 0;
@@ -7180,1757 +6974,10 @@ public class Main extends data
         who_is_playing();
     }
 
-    public void who_is_playing() 
-    {
-        SwingWorker play_win = new SwingWorker()
-        {
-            @Override
-            protected Object doInBackground() throws Exception
-            {
-                Thread.sleep(1000);
-                load.setText("YOU WIN");
-                loading.remove(pb);
-                crd.show(co, "LA");
-                return null;
-            }
-        };
-        SwingWorker com_win = new SwingWorker()
-        {
-            @Override
-            protected Object doInBackground() throws Exception
-            {
-                Thread.sleep(1000);
-                load.setText("YOU LOSE");
-                loading.remove(pb);
-                crd.show(co, "LA");
-                return null;
-            }
-        };
-        if(play_hit_count == 17)
-        {
-            refresh_flag = 2;
-            turn_flag = 2;
-            play_win.execute();
-        }
-        if(com_hit_count == 17)
-        {
-            refresh_flag = 2;
-            turn_flag = 2;
-            com_win.execute();
-        }
-        SwingWorker help = new SwingWorker()
-        {
-            @Override
-            protected String doInBackground() throws Exception
-            {
-                pb.setMaximum(100);
-                pb.setMinimum(0);
-                crd.show(co, "LA");
-                loading.revalidate();
-                loading.repaint();
-                for(int percent = 0 ; percent <= 100 ; percent++)
-                {
-                    pb.setValue(percent);
-                    Thread.sleep(16);
-                }
-                return null;
-            }
-            @Override
-            protected void done()
-            {
-                if(turn_flag == 0)
-                {
-                    crd.show(co, "PHA");
-                }
-                else
-                {
-                    crd.show(co, "CHA");
-                    com_hit();
-                }
-            }
-        };
-        SwingWorker wait = new SwingWorker()
-        {
-            @Override
-            protected Object doInBackground() throws Exception
-            {
-                Thread.sleep(2000);
-                return null;
-            }
-            @Override
-            protected void done()
-            {
-                help.execute();
-            }
-        };
-        hud_change();
-        if(refresh_flag == 1)
-        {
-            wait.execute();
-        }
-        if(refresh_flag == 0 && turn_flag == 1)
-        {
-            com_hit();
-        }
-    }
-    void convert_player()
-    {
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            play_aircraft_pos[i] = Integer.parseInt(aircraft_pos[i].substring(1));
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            play_battleship_pos[i] = Integer.parseInt(battleship_pos[i].substring(1));
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            play_submarine_pos[i] = Integer.parseInt(submarine_pos[i].substring(1));
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            play_cruiser_pos[i] = Integer.parseInt(cruiser_pos[i].substring(1));
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            play_destroyer_pos[i] = Integer.parseInt(destroyer_pos[i].substring(1));
-        }
-        if((play_aircraft_pos[1] - play_aircraft_pos[0]) == 1)
-        {
-            air_or = 0;
-        }
-        else
-        {
-            air_or = 1;
-        }
-        if((play_battleship_pos[1] - play_battleship_pos[0]) == 1)
-        {
-            bat_or = 0;
-        }
-        else
-        {
-            bat_or = 1;
-        }
-        if((play_submarine_pos[1] - play_submarine_pos[0]) == 1)
-        {
-            sub_or = 0;
-        }
-        else
-        {
-            sub_or = 1;
-        }
-        if((play_cruiser_pos[1] - play_cruiser_pos[0]) == 1)
-        {
-            cru_or = 0;
-        }
-        else
-        {
-            cru_or = 1;
-        }
-        if((play_destroyer_pos[1] - play_destroyer_pos[0]) == 1)
-        {
-            des_or = 0;
-        }
-        else
-        {
-            des_or = 1;
-        }
-        System.out.println("Aircraft Carrier Orientation : "+air_or);
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            System.out.println(play_aircraft_pos[i]);
-        }
-        System.out.println("Battleship Orientation : "+bat_or);
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            System.out.println(play_battleship_pos[i]);
-        }
-        System.out.println("Submarine Orientation : "+sub_or);
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            System.out.println(play_submarine_pos[i]);
-        }
-        System.out.println("Cruiser Orientation : "+cru_or);
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            System.out.println(play_cruiser_pos[i]);
-        }
-        System.out.println("Destroyer Orientation : "+des_or);
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            System.out.println(play_destroyer_pos[i]);
-        }
-    }
-    public JButton com_render(JButton b, int pos)
-    {
-        int no_hits = 0;
-        if(air_or == 0)
-        {
-            System.out.println("In Aircraft 0");
-            for(int i = 0 ; i < 5 ; i ++)
-            {
-                if(play_aircraft_pos[i] == pos)
-                {
-                    next_move_flag = "aircraft-carrier";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("aircraft-carrier/aircraft-h-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("aircraft-carrier/aircraft-h-destroyed-tile-"+iconpos+".png"));
-                    com_hit_count++;
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_ai_count++;
-                    no_hits = 1;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(air_or == 1)
-        {
-            System.out.println("In Aircraft 1");
-            for(int i = 0 ; i < 5 ; i ++)
-            {
-                if(play_aircraft_pos[i] == pos)
-                {
-                    next_move_flag = "aircraft-carrier";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_ai_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(bat_or == 0)
-        {
-            System.out.println("In Battleship 0");
-            for(int i = 0 ; i < 4 ; i ++)
-            {
-                if(play_battleship_pos[i] == pos)
-                {
-                    next_move_flag = "battleship";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("battleship/battleship-h-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("battleship/battleship-h-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_bat_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(bat_or == 1)
-        {
-            System.out.println("In Battleship 1");
-            for(int i = 0 ; i < 4 ; i ++)
-            {
-                if(play_battleship_pos[i] == pos)
-                {
-                    next_move_flag = "battleship";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("battleship/battleship-v-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("battleship/battleship-v-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_bat_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(sub_or == 0)
-        {
-            System.out.println("In Submarine 0");
-            for(int i = 0 ; i < 3 ; i ++)
-            {
-                if(play_submarine_pos[i] == pos)
-                {
-                    next_move_flag = "submarine";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("submarine/submarine-h-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("submarine/submarine-h-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_sub_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(sub_or == 1)
-        {
-            System.out.println("In Submarine 1");
-            for(int i = 0 ; i < 3 ; i ++)
-            {
-                if(play_submarine_pos[i] == pos)
-                {
-                    next_move_flag = "submarine";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("submarine/submarine-v-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("submarine/submarine-v-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_sub_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(cru_or == 0)
-        {
-            System.out.println("In Cruiser 0");
-            for(int i = 0 ; i < 3 ; i ++)
-            {
-                if(play_cruiser_pos[i] == pos)
-                {
-                    next_move_flag = "cruiser";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("cruiser/cruiser-h-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("cruiser/cruiser-h-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_cru_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(cru_or == 1)
-        {
-            System.out.println("In Cruiser 1");
-            for(int i = 0 ; i < 3 ; i ++)
-            {
-                if(play_cruiser_pos[i] == pos)
-                {
-                    next_move_flag = "cruiser";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("cruiser/cruiser-v-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("cruiser/cruiser-v-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_cru_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(des_or == 0)
-        {
-            System.out.println("In Destroyer 0");
-            for(int i = 0 ; i < 2 ; i ++)
-            {
-                if(play_destroyer_pos[i] == pos)
-                {
-                    next_move_flag = "destroyer";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("destroyer/destroyer-h-destroyed-tile"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("destroyer/destroyer-h-destroyed-tile"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_des_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(des_or == 1)
-        {
-            System.out.println("In Destroyer 1");
-            for(int i = 0 ; i < 2 ; i ++)
-            {
-                if(play_destroyer_pos[i] == pos)
-                {
-                    next_move_flag = "destroyer";
-                    int iconpos = i + 1;
-                    com_is_hit_or_not = 1;
-                    b.setIcon(new ImageIcon("destroyer/destroyer-v-destroyed-tile-"+iconpos+".png"));
-                    b.setDisabledIcon(new ImageIcon("destroyer/destroyer-v-destroyed-tile-"+iconpos+".png"));
-                    System.out.println("aircraft-carrier/aircraft-v-destroyed-tile-"+iconpos+".png");
-                    com_des_count++;
-                    no_hits = 1;
-                    com_hit_count++;
-                    j.revalidate();
-                    j.repaint();
-                    return b;
-                }
-            }
-        }
-        if(no_hits == 0)
-        {
-            com_is_hit_or_not = 0;
-            b.setIcon(new ImageIcon("water_ship_miss.png"));
-            miss();
-            j.revalidate();
-            j.repaint();
-            return b;
-        }
-        return b;
-    }
-    public void com_hit()
-    {
-        SwingWorker com_hitter = new SwingWorker()
-        {
-            @Override
-            protected Object doInBackground() throws Exception
-            {
-                int loop_count = 0;
-                while(true)
-                {
-                    int hit_pos_index;
-                    if(next_move == -1)
-                    {
-                        hit_pos_index = Math.abs(r.nextInt());
-                        hit_pos_index = hit_pos_index % 100;
-                    }
-                    else
-                    {
-                        hit_pos_index = next_move;
-                    }
-                    System.out.println(hit_pos_index);
-                    if(com_hit_grid[hit_pos_index] != 6000)
-                    {
-                        com_hit_grid[hit_pos_index] = 6000;
-                        if(hit_pos_index == 0)
-                        {
-                            j1 = com_render(j1, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 1)
-                        {
-                            j2 = com_render(j2, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 2)
-                        {
-                            j3 = com_render(j3, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 3)
-                        {
-                            j4 = com_render(j4, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 4)
-                        {
-                            j5 = com_render(j5, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 5)
-                        {
-                            j6 = com_render(j6, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 6)
-                        {
-                            j7 = com_render(j7, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 7)
-                        {
-                            j8 = com_render(j8, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 8)
-                        {
-                            j9 = com_render(j9, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 9)
-                        {
-                            j10 = com_render(j10, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 10)
-                        {
-                            j11 = com_render(j11, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 11)
-                        {
-                            j12 = com_render(j12, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 12)
-                        {
-                            j13 = com_render(j13, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 13)
-                        {
-                            j14 = com_render(j14, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 14)
-                        {
-                            j15 = com_render(j15, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 15)
-                        {
-                            j16 = com_render(j16, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 16)
-                        {
-                            j17 = com_render(j17, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 17)
-                        {
-                            j18 = com_render(j18, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 18)
-                        {
-                            j19 = com_render(j19, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 19)
-                        {
-                            j20 = com_render(j20, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 20)
-                        {
-                            j21 = com_render(j21, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 21)
-                        {
-                            j22 = com_render(j22, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 22)
-                        {
-                            j23 = com_render(j23, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 23)
-                        {
-                            j24 = com_render(j24, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 24)
-                        {
-                            j25 = com_render(j25, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 25)
-                        {
-                            j26 = com_render(j26, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 26)
-                        {
-                            j27 = com_render(j27, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 27)
-                        {
-                            j28 = com_render(j28, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 28)
-                        {
-                            j29 = com_render(j29, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 29)
-                        {
-                            j30 = com_render(j30, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 30)
-                        {
-                            j31 = com_render(j31, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 31)
-                        {
-                            j32 = com_render(j32, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 32)
-                        {
-                            j33 = com_render(j33, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 33)
-                        {
-                            j34 = com_render(j34, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 34)
-                        {
-                            j35 = com_render(j35, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 35)
-                        {
-                            j36 = com_render(j36, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 36)
-                        {
-                            j37 = com_render(j37, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 37)
-                        {
-                            j38 = com_render(j38, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 38)
-                        {
-                            j39 = com_render(j39, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 39)
-                        {
-                            j40 = com_render(j40, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 40)
-                        {
-                            j41 = com_render(j41, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 41)
-                        {
-                            j42 = com_render(j42, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 42)
-                        {
-                            j43 = com_render(j43, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 43)
-                        {
-                            j44 = com_render(j44, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 44)
-                        {
-                            j45 = com_render(j45, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 45)
-                        {
-                            j46 = com_render(j46, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 46)
-                        {
-                            j47 = com_render(j47, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 47)
-                        {
-                            j48 = com_render(j48, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 48)
-                        {
-                            j49 = com_render(j49, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 49)
-                        {
-                            j50 = com_render(j50, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 50)
-                        {
-                            j51 = com_render(j51, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 51)
-                        {
-                            j52 = com_render(j52, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 52)
-                        {
-                            j53 = com_render(j53, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 53)
-                        {
-                            j54 = com_render(j54, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 54)
-                        {
-                            j55 = com_render(j55, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 55)
-                        {
-                            j56 = com_render(j56, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 56)
-                        {
-                            j57 = com_render(j57, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 57)
-                        {
-                            j58 = com_render(j58, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 58)
-                        {
-                            j59 = com_render(j59, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 59)
-                        {
-                            j60 = com_render(j60, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 60)
-                        {
-                            j61 = com_render(j61, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 61)
-                        {
-                            j62 = com_render(j62, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 62)
-                        {
-                            j63 = com_render(j63, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 63)
-                        {
-                            j64 = com_render(j64, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 64)
-                        {
-                            j65 = com_render(j65, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 65)
-                        {
-                            j66 = com_render(j66, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 66)
-                        {
-                            j67 = com_render(j67, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 67)
-                        {
-                            j68 = com_render(j68, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 68)
-                        {
-                            j69 = com_render(j69, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 69)
-                        {
-                            j70 = com_render(j70, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 70)
-                        {
-                            j71 = com_render(j71, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 71)
-                        {
-                            j72 = com_render(j72, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 72)
-                        {
-                            j73 = com_render(j73, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 73)
-                        {
-                            j74 = com_render(j74, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 74)
-                        {
-                            j75 = com_render(j75, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 75)
-                        {
-                            j76 = com_render(j76, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 76)
-                        {
-                            j77 = com_render(j77, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 77)
-                        {
-                            j78 = com_render(j78, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 78)
-                        {
-                            j79 = com_render(j79, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 79)
-                        {
-                            j80 = com_render(j80, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 80)
-                        {
-                            j81 = com_render(j81, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 81)
-                        {
-                            j82 = com_render(j82, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 82)
-                        {
-                            j83 = com_render(j83, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 83)
-                        {
-                            j84 = com_render(j84, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 84)
-                        {
-                            j85 = com_render(j85, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 85)
-                        {
-                            j86 = com_render(j86, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 86)
-                        {
-                            j87 = com_render(j87, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 87)
-                        {
-                            j88 = com_render(j88, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 88)
-                        {
-                            j89 = com_render(j89, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 89)
-                        {
-                            j90 = com_render(j90, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 90)
-                        {
-                            j91 = com_render(j91, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 91)
-                        {
-                            j92 = com_render(j92, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 92)
-                        {
-                            j93 = com_render(j93, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 93)
-                        {
-                            j94 = com_render(j94, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 94)
-                        {
-                            j95 = com_render(j95, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 95)
-                        {
-                            j96 = com_render(j96, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 96)
-                        {
-                            j97 = com_render(j97, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 97)
-                        {
-                            j98 = com_render(j98, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 98)
-                        {
-                            j99 = com_render(j99, hit_pos_index+1);
-                        }
-                        else if(hit_pos_index == 99)
-                        {
-                            j100 = com_render(j100, hit_pos_index+1);
-                        }
-                        j.revalidate();
-                        j.repaint();
-                        if(com_is_hit_or_not == 0)
-                        {
-                            refresh_flag = 1;
-                            turn_flag = 0;
-                        }
-                        else
-                        {
-                            refresh_flag = 0;
-                            turn_flag = 1;
-                        }
-                        if(Objects.equals(next_move_flag, "aircraft-carrier"))
-                        {
-                            if(com_ai_count == 1)
-                            {
-                                if (air_already_assigned == 0)
-                                {
-                                    System.out.println("hit_pos_index = " + hit_pos_index);
-                                    air_already_assigned = 1;
-                                    up_move = hit_pos_index - 10;
-                                    down_move = hit_pos_index + 10;
-                                    left_move = hit_pos_index - 1;
-                                    right_move = hit_pos_index + 1;
-                                    if (up_move < 0 || com_hit_grid[up_move] == 6000)
-                                    {
-                                        up_move = -1;
-                                    }
-                                    if (down_move >= 100 || com_hit_grid[down_move] == 6000)
-                                    {
-                                        down_move = -1;
-                                    }
-                                    if ((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000)
-                                    {
-                                        left_move = -1;
-                                    }
-                                    if (right_move % 10 == 0 || com_hit_grid[right_move] == 6000)
-                                    {
-                                        right_move = -1;
-                                    }
-                                }
-                                if(up_move != -1)
-                                {
-                                    next_move = up_move;
-                                    air_calculated_orientation = "up";
-                                    up_move = -1;
-                                }
-                                else if(down_move != -1)
-                                {
-                                    next_move = down_move;
-                                    air_calculated_orientation = "down";
-                                    down_move = -1;
-                                }
-                                else if(left_move != -1)
-                                {
-                                    next_move = left_move;
-                                    air_calculated_orientation = "left";
-                                    left_move = -1;
-                                }
-                                else if(right_move != -1)
-                                {
-                                    next_move = right_move;
-                                    air_calculated_orientation = "right";
-                                    right_move = -1;
-                                }
-                            }
-                            if(com_ai_count == 2)
-                            {
-                                if(Objects.equals(air_calculated_orientation, "up"))
-                                {
-                                    if(air_up_assigned_1 == 0)
-                                    {
-                                        air_up_assigned_1 = 1;
-                                        up_move = next_move - 10;
-                                        other_move = down_move;
-                                    }
-                                    if(up_move < 0 || com_hit_grid[up_move] == 6000 || up_move == -1)
-                                    {
-                                        up_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        air_calculated_orientation = "down";
-                                    }
-                                    else
-                                    {
-                                        next_move = up_move;
-                                        up_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "down"))
-                                {
-                                    if(air_down_assigned_1 == 0)
-                                    {
-                                        air_down_assigned_1 = 1;
-                                        down_move = next_move + 10;
-                                    }
-                                    if(down_move >= 100 || com_hit_grid[down_move] == 6000 || down_move == -1)
-                                    {
-                                        down_move = -1;
-                                        next_move = down_move;
-                                        air_calculated_orientation = "up";
-                                    }
-                                    else
-                                    {
-                                        next_move = down_move;
-                                        down_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "left"))
-                                {
-                                    if(air_left_assigned_1 == 0)
-                                    {
-                                        air_left_assigned_1 = 1;
-                                        left_move = next_move - 1;
-                                        other_move = right_move;
-                                    }
-                                    if((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000 || left_move  == -1)
-                                    {
-                                        left_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        air_calculated_orientation = "right";
-                                    }
-                                    else
-                                    {
-                                        next_move = left_move;
-                                        left_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "right"))
-                                {
-                                    if(air_right_assigned_1 == 0)
-                                    {
-                                        air_right_assigned_1 = 1;
-                                        right_move = next_move + 1;
-                                    }
-                                    if(right_move % 10 == 0 || com_hit_grid[right_move] == 6000 || right_move == -1)
-                                    {
-                                        right_move = -1;
-                                        next_move = right_move;
-                                        air_calculated_orientation = "left";
-                                    }
-                                    else
-                                    {
-                                        next_move = right_move;
-                                        right_move = -1;
-                                    }
-                                }
-                            }
-                            if(com_ai_count == 3)
-                            {
-                                if(Objects.equals(air_calculated_orientation, "up"))
-                                {
-                                    if(air_up_assigned_2 == 0)
-                                    {
-                                        air_up_assigned_2 = 1;
-                                        up_move = next_move - 10;
-                                    }
-                                    if(up_move < 0 || com_hit_grid[up_move] == 6000 || up_move == -1)
-                                    {
-                                        up_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        air_calculated_orientation = "down";
-                                    }
-                                    else
-                                    {
-                                        next_move = up_move;
-                                        up_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "down"))
-                                {
-                                    if(air_down_assigned_2 == 0)
-                                    {
-                                        air_down_assigned_2 = 1;
-                                        down_move = next_move + 10;
-                                    }
-                                    if(down_move >= 100 || com_hit_grid[down_move] == 6000 || down_move == -1)
-                                    {
-                                        down_move = -1;
-                                        next_move = down_move;
-                                        air_calculated_orientation = "up";
-                                    }
-                                    else
-                                    {
-                                        next_move = down_move;
-                                        down_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "left"))
-                                {
-                                    if(air_left_assigned_2 == 0)
-                                    {
-                                        air_left_assigned_2 = 1;
-                                        left_move = next_move - 1;
-                                    }
-                                    if((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000 || left_move  == -1)
-                                    {
-                                        left_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        air_calculated_orientation = "right";
-                                    }
-                                    else
-                                    {
-                                        next_move = left_move;
-                                        left_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "right"))
-                                {
-                                    if(air_right_assigned_2 == 0)
-                                    {
-                                        air_right_assigned_2 = 1;
-                                        right_move = next_move + 1;
-                                    }
-                                    if(right_move % 10 == 0 || com_hit_grid[right_move] == 6000 || right_move == -1)
-                                    {
-                                        right_move = -1;
-                                        next_move = right_move;
-                                        air_calculated_orientation = "left";
-                                    }
-                                    else
-                                    {
-                                        next_move = right_move;
-                                        right_move = -1;
-                                    }
-                                }
-                            }
-                            if(com_ai_count == 4)
-                            {
-                                if(Objects.equals(air_calculated_orientation, "up"))
-                                {
-                                    if(air_up_assigned_3 == 0)
-                                    {
-                                        air_up_assigned_3 = 1;
-                                        up_move = next_move - 10;
-                                    }
-                                    if(up_move < 0 || com_hit_grid[up_move] == 6000 || up_move == -1)
-                                    {
-                                        up_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        air_calculated_orientation = "down";
-                                    }
-                                    else
-                                    {
-                                        next_move = up_move;
-                                        up_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "down"))
-                                {
-                                    if(air_down_assigned_3 == 0)
-                                    {
-                                        air_down_assigned_3 = 1;
-                                        down_move = next_move + 10;
-                                    }
-                                    if(down_move >= 100 || com_hit_grid[down_move] == 6000 || down_move == -1)
-                                    {
-                                        down_move = -1;
-                                        next_move = down_move;
-                                        air_calculated_orientation = "up";
-                                    }
-                                    else
-                                    {
-                                        next_move = down_move;
-                                        down_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "left"))
-                                {
-                                    if(air_left_assigned_3 == 0)
-                                    {
-                                        air_left_assigned_3 = 1;
-                                        left_move = next_move - 1;
-                                    }
-                                    if((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000 || left_move  == -1)
-                                    {
-                                        left_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        air_calculated_orientation = "right";
-                                    }
-                                    else
-                                    {
-                                        next_move = left_move;
-                                        left_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(air_calculated_orientation, "right"))
-                                {
-                                    if(air_right_assigned_3 == 0)
-                                    {
-                                        air_right_assigned_3 = 1;
-                                        right_move = next_move + 1;
-                                    }
-                                    if(right_move % 10 == 0 || com_hit_grid[right_move] == 6000 || right_move == -1)
-                                    {
-                                        right_move = -1;
-                                        next_move = right_move;
-                                        air_calculated_orientation = "left";
-                                    }
-                                    else
-                                    {
-                                        next_move = right_move;
-                                        right_move = -1;
-                                    }
-                                }
-                            }
-                            if(com_ai_count == 5)
-                            {
-                                next_move_flag = "none";
-                                next_move = -1;
-                            }
-                        }
-                        else if(Objects.equals(next_move_flag, "battleship"))
-                        {
-                            if(com_bat_count == 1)
-                            {
-                                if (bat_already_assigned == 0)
-                                {
-                                    System.out.println("hit_pos_index = " + hit_pos_index);
-                                    bat_already_assigned = 1;
-                                    up_move = hit_pos_index - 10;
-                                    down_move = hit_pos_index + 10;
-                                    left_move = hit_pos_index - 1;
-                                    right_move = hit_pos_index + 1;
-                                    if (up_move < 0 || com_hit_grid[up_move] == 6000)
-                                    {
-                                        up_move = -1;
-                                    }
-                                    if (down_move >= 100 || com_hit_grid[down_move] == 6000)
-                                    {
-                                        down_move = -1;
-                                    }
-                                    if ((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000)
-                                    {
-                                        left_move = -1;
-                                    }
-                                    if (right_move % 10 == 0 || com_hit_grid[right_move] == 6000)
-                                    {
-                                        right_move = -1;
-                                    }
-                                }
-                                if(up_move != -1)
-                                {
-                                    next_move = up_move;
-                                    bat_calculated_orientation = "up";
-                                    up_move = -1;
-                                }
-                                else if(down_move != -1)
-                                {
-                                    next_move = down_move;
-                                    bat_calculated_orientation = "down";
-                                    down_move = -1;
-                                }
-                                else if(left_move != -1)
-                                {
-                                    next_move = left_move;
-                                    bat_calculated_orientation = "left";
-                                    left_move = -1;
-                                }
-                                else if(right_move != -1)
-                                {
-                                    next_move = right_move;
-                                    bat_calculated_orientation = "right";
-                                    right_move = -1;
-                                }
-                            }
-                            if(com_bat_count == 2)
-                            {
-                                if(Objects.equals(bat_calculated_orientation, "up"))
-                                {
-                                    if(bat_up_assigned_1 == 0)
-                                    {
-                                        bat_up_assigned_1 = 1;
-                                        up_move = next_move - 10;
-                                        other_move = down_move;
-                                    }
-                                    if(up_move < 0 || com_hit_grid[up_move] == 6000 || up_move == -1)
-                                    {
-                                        up_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        bat_calculated_orientation = "down";
-                                    }
-                                    else
-                                    {
-                                        next_move = up_move;
-                                        up_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(bat_calculated_orientation, "down"))
-                                {
-                                    if(bat_down_assigned_1 == 0)
-                                    {
-                                        bat_down_assigned_1 = 1;
-                                        down_move = next_move + 10;
-                                    }
-                                    if(down_move >= 100 || com_hit_grid[down_move] == 6000 || down_move == -1)
-                                    {
-                                        down_move = -1;
-                                        next_move = down_move;
-                                        bat_calculated_orientation = "up";
-                                    }
-                                    else
-                                    {
-                                        next_move = down_move;
-                                        down_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(bat_calculated_orientation, "left"))
-                                {
-                                    if(bat_left_assigned_1 == 0)
-                                    {
-                                        bat_left_assigned_1 = 1;
-                                        left_move = next_move - 1;
-                                        other_move = right_move;
-                                    }
-                                    if((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000 || left_move  == -1)
-                                    {
-                                        left_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        bat_calculated_orientation = "right";
-                                    }
-                                    else
-                                    {
-                                        next_move = left_move;
-                                        left_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(bat_calculated_orientation, "right"))
-                                {
-                                    if(bat_right_assigned_1 == 0)
-                                    {
-                                        bat_right_assigned_1 = 1;
-                                        right_move = next_move + 1;
-                                    }
-                                    if(right_move % 10 == 0 || com_hit_grid[right_move] == 6000 || right_move == -1)
-                                    {
-                                        right_move = -1;
-                                        next_move = right_move;
-                                        bat_calculated_orientation = "left";
-                                    }
-                                    else
-                                    {
-                                        next_move = right_move;
-                                        right_move = -1;
-                                    }
-                                }
-                            }
-                            if(com_bat_count == 3)
-                            {
-                                if(Objects.equals(bat_calculated_orientation, "up"))
-                                {
-                                    if(bat_up_assigned_2 == 0)
-                                    {
-                                        bat_up_assigned_2 = 1;
-                                        up_move = next_move - 10;
-                                    }
-                                    if(up_move < 0 || com_hit_grid[up_move] == 6000 || up_move == -1)
-                                    {
-                                        up_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        bat_calculated_orientation = "down";
-                                    }
-                                    else
-                                    {
-                                        next_move = up_move;
-                                        up_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(bat_calculated_orientation, "down"))
-                                {
-                                    if(bat_down_assigned_2 == 0)
-                                    {
-                                        bat_down_assigned_2 = 1;
-                                        down_move = next_move + 10;
-                                    }
-                                    if(down_move >= 100 || com_hit_grid[down_move] == 6000 || down_move == -1)
-                                    {
-                                        down_move = -1;
-                                        next_move = down_move;
-                                        bat_calculated_orientation = "up";
-                                    }
-                                    else
-                                    {
-                                        next_move = down_move;
-                                        down_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(bat_calculated_orientation, "left"))
-                                {
-                                    if(bat_left_assigned_2 == 0)
-                                    {
-                                        bat_left_assigned_2 = 1;
-                                        left_move = next_move - 1;
-                                    }
-                                    if((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000 || left_move  == -1)
-                                    {
-                                        left_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        bat_calculated_orientation = "right";
-                                    }
-                                    else
-                                    {
-                                        next_move = left_move;
-                                        left_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(bat_calculated_orientation, "right"))
-                                {
-                                    if(bat_right_assigned_2 == 0)
-                                    {
-                                        bat_right_assigned_2 = 1;
-                                        right_move = next_move + 1;
-                                    }
-                                    if(right_move % 10 == 0 || com_hit_grid[right_move] == 6000 || right_move == -1)
-                                    {
-                                        right_move = -1;
-                                        next_move = right_move;
-                                        bat_calculated_orientation = "left";
-                                    }
-                                    else
-                                    {
-                                        next_move = right_move;
-                                        right_move = -1;
-                                    }
-                                }
-                            }
-                            if(com_bat_count == 4)
-                            {
-                                next_move_flag = "none";
-                                next_move = -1;
-                            }
-                        }
-                        else if(Objects.equals(next_move_flag, "submarine"))
-                        {
-                            if(com_sub_count == 1)
-                            {
-                                if (sub_already_assigned == 0)
-                                {
-                                    sub_already_assigned = 1;
-                                    up_move = hit_pos_index - 10;
-                                    down_move = hit_pos_index + 10;
-                                    left_move = hit_pos_index - 1;
-                                    right_move = hit_pos_index + 1;
-                                    if (up_move < 0 || com_hit_grid[up_move] == 6000)
-                                    {
-                                        up_move = -1;
-                                    }
-                                    if (down_move >= 100 || com_hit_grid[down_move] == 6000)
-                                    {
-                                        down_move = -1;
-                                    }
-                                    if ((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000)
-                                    {
-                                        left_move = -1;
-                                    }
-                                    if (right_move % 10 == 0 || com_hit_grid[right_move] == 6000)
-                                    {
-                                        right_move = -1;
-                                    }
-                                }
-                                if(up_move != -1)
-                                {
-                                    next_move = up_move;
-                                    sub_calculated_orientation = "up";
-                                    up_move = -1;
-                                }
-                                else if(down_move != -1)
-                                {
-                                    next_move = down_move;
-                                    sub_calculated_orientation = "down";
-                                    down_move = -1;
-                                }
-                                else if(left_move != -1)
-                                {
-                                    next_move = left_move;
-                                    sub_calculated_orientation = "left";
-                                    left_move = -1;
-                                }
-                                else if(right_move != -1)
-                                {
-                                    next_move = right_move;
-                                    sub_calculated_orientation = "right";
-                                    right_move = -1;
-                                }
-                            }
-                            if(com_sub_count == 2)
-                            {
-                                if(Objects.equals(sub_calculated_orientation, "up"))
-                                {
-                                    if(sub_up_assigned_1 == 0)
-                                    {
-                                        sub_up_assigned_1 = 1;
-                                        up_move = next_move - 10;
-                                        other_move = down_move;
-                                        System.out.println("Other Move = " + other_move);
-                                    }
-                                    if(up_move < 0 || com_hit_grid[up_move] == 6000 || up_move == -1)
-                                    {
-                                        up_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        sub_calculated_orientation = "down";
-                                    }
-                                    else
-                                    {
-                                        next_move = up_move;
-                                        up_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(sub_calculated_orientation, "down"))
-                                {
-                                    if(sub_down_assigned_1 == 0)
-                                    {
-                                        sub_down_assigned_1 = 1;
-                                        down_move = next_move + 10;
-                                    }
-                                    if(down_move >= 100 || com_hit_grid[down_move] == 6000 || down_move == -1)
-                                    {
-                                        down_move = -1;
-                                        next_move = down_move;
-                                        sub_calculated_orientation = "up";
-                                    }
-                                    else
-                                    {
-                                        next_move = down_move;
-                                        down_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(sub_calculated_orientation, "left"))
-                                {
-                                    if(sub_left_assigned_1 == 0)
-                                    {
-                                        sub_left_assigned_1 = 1;
-                                        left_move = next_move - 1;
-                                        other_move = right_move;
-                                        System.out.println("Other Move = " + other_move);
-                                    }
-                                    if((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000 || left_move  == -1)
-                                    {
-                                        left_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        sub_calculated_orientation = "right";
-                                    }
-                                    else
-                                    {
-                                        next_move = left_move;
-                                        left_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(sub_calculated_orientation, "right"))
-                                {
-                                    if(sub_right_assigned_1 == 0)
-                                    {
-                                        sub_right_assigned_1 = 1;
-                                        right_move = next_move + 1;
-                                    }
-                                    if(right_move % 10 == 0 || com_hit_grid[right_move] == 6000 || right_move == -1)
-                                    {
-                                        right_move = -1;
-                                        next_move = right_move;
-                                        sub_calculated_orientation = "left";
-                                    }
-                                    else
-                                    {
-                                        next_move = right_move;
-                                        right_move = -1;
-                                    }
-                                }
-                            }
-                            if(com_sub_count == 3)
-                            {
-                                next_move_flag = "none";
-                                next_move = -1;
-                            }
-                        }
-                        else if(Objects.equals(next_move_flag, "cruiser"))
-                        {
-                            if(com_cru_count == 1)
-                            {
-                                if (cru_already_assigned == 0)
-                                {
-                                    cru_already_assigned = 1;
-                                    up_move = hit_pos_index - 10;
-                                    down_move = hit_pos_index + 10;
-                                    left_move = hit_pos_index - 1;
-                                    right_move = hit_pos_index + 1;
-                                    if (up_move < 0 || com_hit_grid[up_move] == 6000)
-                                    {
-                                        up_move = -1;
-                                    }
-                                    if (down_move >= 100 || com_hit_grid[down_move] == 6000)
-                                    {
-                                        down_move = -1;
-                                    }
-                                    if ((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000)
-                                    {
-                                        left_move = -1;
-                                    }
-                                    if (right_move % 10 == 0 || com_hit_grid[right_move] == 6000)
-                                    {
-                                        right_move = -1;
-                                    }
-                                }
-                                if(up_move != -1)
-                                {
-                                    next_move = up_move;
-                                    cru_calculated_orientation = "up";
-                                    up_move = -1;
-                                }
-                                else if(down_move != -1)
-                                {
-                                    next_move = down_move;
-                                    cru_calculated_orientation = "down";
-                                    down_move = -1;
-                                }
-                                else if(left_move != -1)
-                                {
-                                    next_move = left_move;
-                                    cru_calculated_orientation = "left";
-                                    left_move = -1;
-                                }
-                                else if(right_move != -1)
-                                {
-                                    next_move = right_move;
-                                    cru_calculated_orientation = "right";
-                                    right_move = -1;
-                                }
-                            }
-                            if(com_cru_count == 2)
-                            {
-                                if(Objects.equals(cru_calculated_orientation, "up"))
-                                {
-                                    if(cru_up_assigned_1 == 0)
-                                    {
-                                        cru_up_assigned_1 = 1;
-                                        up_move = next_move - 10;
-                                        other_move = down_move;
-                                    }
-                                    if(up_move < 0 || com_hit_grid[up_move] == 6000 || up_move == -1)
-                                    {
-                                        up_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        cru_calculated_orientation = "down";
-                                    }
-                                    else
-                                    {
-                                        next_move = up_move;
-                                        up_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(cru_calculated_orientation, "down"))
-                                {
-                                    if(cru_down_assigned_1 == 0)
-                                    {
-                                        cru_down_assigned_1 = 1;
-                                        down_move = next_move + 10;
-                                    }
-                                    if(down_move >= 100 || com_hit_grid[down_move] == 6000 || down_move == -1)
-                                    {
-                                        down_move = -1;
-                                        next_move = down_move;
-                                        cru_calculated_orientation = "up";
-                                    }
-                                    else
-                                    {
-                                        next_move = down_move;
-                                        down_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(cru_calculated_orientation, "left"))
-                                {
-                                    if(cru_left_assigned_1 == 0)
-                                    {
-                                        cru_left_assigned_1 = 1;
-                                        left_move = next_move - 1;
-                                        other_move = right_move;
-                                    }
-                                    if((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000 || left_move  == -1)
-                                    {
-                                        left_move = -1;
-                                        next_move = other_move;
-                                        other_move = -1;
-                                        cru_calculated_orientation = "right";
-                                    }
-                                    else
-                                    {
-                                        next_move = left_move;
-                                        left_move = -1;
-                                    }
-                                }
-                                else if(Objects.equals(cru_calculated_orientation, "right"))
-                                {
-                                    if(cru_right_assigned_1 == 0)
-                                    {
-                                        cru_right_assigned_1 = 1;
-                                        right_move = next_move + 1;
-                                    }
-                                    if(right_move % 10 == 0 || com_hit_grid[right_move] == 6000 || right_move == -1)
-                                    {
-                                        right_move = -1;
-                                        next_move = right_move;
-                                        cru_calculated_orientation = "left";
-                                    }
-                                    else
-                                    {
-                                        next_move = right_move;
-                                        right_move = -1;
-                                    }
-                                }
-                            }
-                            if(com_cru_count == 3)
-                            {
-                                next_move_flag = "none";
-                                next_move = -1;
-                            }
-                        }
-                        else if(Objects.equals(next_move_flag, "destroyer"))
-                        {
-                            if(des_already_assigned == 0)
-                            {
-                                des_already_assigned = 1;
-                                up_move = hit_pos_index - 10;
-                                down_move = hit_pos_index + 10;
-                                left_move = hit_pos_index - 1;
-                                right_move = hit_pos_index + 1;
-                                if (up_move < 0 || com_hit_grid[up_move] == 6000)
-                                {
-                                    up_move = -1;
-                                }
-                                if (down_move >= 100 || com_hit_grid[down_move] == 6000)
-                                {
-                                    down_move = -1;
-                                }
-                                if ((left_move + 1) % 10 == 0 || com_hit_grid[left_move] == 6000)
-                                {
-                                    left_move = -1;
-                                }
-                                if (right_move % 10 == 0 || com_hit_grid[right_move] == 6000)
-                                {
-                                    right_move = -1;
-                                }
-                            }
-                            if(up_move != -1)
-                            {
-                                next_move = up_move;
-                                up_move = -1;
-                            }
-                            else if(down_move != -1)
-                            {
-                                next_move = down_move;
-                                down_move = -1;
-                            }
-                            else if(left_move != -1)
-                            {
-                                next_move = left_move;
-                                left_move = -1;
-                            }
-                            else if(right_move != -1)
-                            {
-                                next_move = right_move;
-                                right_move = -1;
-                            }
-                            if(com_des_count == 2)
-                            {
-                                next_move_flag = "none";
-                                next_move = -1;
-                            }
-                        }
-                        break;
-                    }
-                    loop_count++;
-                    if(loop_count == 200)
-                    {
-                        next_move = -1;
-                        next_move_flag = "none";
-                    }
-                }
-                return null;
-            }
-            @Override
-            protected void done()
-            {
-                who_is_playing();
-            }
-        };
-        SwingWorker com_delay = new SwingWorker()
-        {
-            @Override
-            protected Object doInBackground() throws Exception
-            {
-                Thread.sleep(1000);
-                return null;
-            }
-            @Override
-            protected void done()
-            {
-                com_hitter.execute();
-            }
-        };
-        com_delay.execute();
-    }
+
+
+
+
     public void j1Call()
     {
         j.remove(l);
@@ -31439,1460 +29486,7 @@ public class Main extends data
         loading.revalidate();
         loading.repaint();
     }
-    public void com_place()
-    {
-        j.remove(tip);
-        convert_player();
-        Icon i = new ImageIcon("water.png");
-        l1 = new JButton(i);
-        l2 = new JButton(i);
-        l3 = new JButton(i);
-        l4 = new JButton(i);
-        l5 = new JButton(i);
-        l6 = new JButton(i);
-        l7 = new JButton(i);
-        l8 = new JButton(i);
-        l9 = new JButton(i);
-        l10 = new JButton(i);
-        l11 = new JButton(i);
-        l12 = new JButton(i);
-        l13 = new JButton(i);
-        l14 = new JButton(i);
-        l15 = new JButton(i);
-        l16 = new JButton(i);
-        l17 = new JButton(i);
-        l18 = new JButton(i);
-        l19 = new JButton(i);
-        l20 = new JButton(i);
-        l21 = new JButton(i);
-        l22 = new JButton(i);
-        l23 = new JButton(i);
-        l24 = new JButton(i);
-        l25 = new JButton(i);
-        l26 = new JButton(i);
-        l27 = new JButton(i);
-        l28 = new JButton(i);
-        l29 = new JButton(i);
-        l30 = new JButton(i);
-        l31 = new JButton(i);
-        l32 = new JButton(i);
-        l33 = new JButton(i);
-        l34 = new JButton(i);
-        l35 = new JButton(i);
-        l36 = new JButton(i);
-        l37 = new JButton(i);
-        l38 = new JButton(i);
-        l39 = new JButton(i);
-        l40 = new JButton(i);
-        l41 = new JButton(i);
-        l42 = new JButton(i);
-        l43 = new JButton(i);
-        l44 = new JButton(i);
-        l45 = new JButton(i);
-        l46 = new JButton(i);
-        l47 = new JButton(i);
-        l48 = new JButton(i);
-        l49 = new JButton(i);
-        l50 = new JButton(i);
-        l51 = new JButton(i);
-        l52 = new JButton(i);
-        l53 = new JButton(i);
-        l54 = new JButton(i);
-        l55 = new JButton(i);
-        l56 = new JButton(i);
-        l57 = new JButton(i);
-        l58 = new JButton(i);
-        l59 = new JButton(i);
-        l60 = new JButton(i);
-        l61 = new JButton(i);
-        l62 = new JButton(i);
-        l63 = new JButton(i);
-        l64 = new JButton(i);
-        l65 = new JButton(i);
-        l66 = new JButton(i);
-        l67 = new JButton(i);
-        l68 = new JButton(i);
-        l69 = new JButton(i);
-        l70 = new JButton(i);
-        l71 = new JButton(i);
-        l72 = new JButton(i);
-        l73 = new JButton(i);
-        l74 = new JButton(i);
-        l75 = new JButton(i);
-        l76 = new JButton(i);
-        l77 = new JButton(i);
-        l78 = new JButton(i);
-        l79 = new JButton(i);
-        l80 = new JButton(i);
-        l81 = new JButton(i);
-        l82 = new JButton(i);
-        l83 = new JButton(i);
-        l84 = new JButton(i);
-        l85 = new JButton(i);
-        l86 = new JButton(i);
-        l87 = new JButton(i);
-        l88 = new JButton(i);
-        l89 = new JButton(i);
-        l90 = new JButton(i);
-        l91 = new JButton(i);
-        l92 = new JButton(i);
-        l93 = new JButton(i);
-        l94 = new JButton(i);
-        l95 = new JButton(i);
-        l96 = new JButton(i);
-        l97 = new JButton(i);
-        l98 = new JButton(i);
-        l99 = new JButton(i);
-        l100 = new JButton(i);
-        tp_air = new JLabel("Aircraft Carrier");
-        tp_bat = new JLabel("Battleship");
-        tp_sub = new JLabel("Submarine");
-        tp_cru = new JLabel("Cruiser");
-        tp_des = new JLabel("Destroyer");
-        tp_air.setForeground(Color.WHITE);
-        tp_bat.setForeground(Color.WHITE);
-        tp_sub.setForeground(Color.WHITE);
-        tp_cru.setForeground(Color.WHITE);
-        tp_des.setForeground(Color.WHITE);
-        p_air = new JLabel(safe);
-        p_bat = new JLabel(safe);
-        p_sub = new JLabel(safe);
-        p_cru = new JLabel(safe);
-        p_des = new JLabel(safe);
-        tc_air = new JLabel("Aircraft Carrier");
-        tc_bat = new JLabel("Battleship");
-        tc_sub = new JLabel("Submarine");
-        tc_cru = new JLabel("Cruiser");
-        tc_des = new JLabel("Destroyer");
-        tc_air.setForeground(Color.WHITE);
-        tc_bat.setForeground(Color.WHITE);
-        tc_sub.setForeground(Color.WHITE);
-        tc_cru.setForeground(Color.WHITE);
-        tc_des.setForeground(Color.WHITE);
-        c_air = new JLabel(hit);
-        c_bat = new JLabel(hit);
-        c_sub = new JLabel(hit);
-        c_cru = new JLabel(hit);
-        c_des = new JLabel(hit);
-        l1.setBounds(150,40,50,50);
-        l2.setBounds(200,40,50,50);
-        l3.setBounds(250,40,50,50);
-        l4.setBounds(300,40,50,50);
-        l5.setBounds(350,40,50,50);
-        l6.setBounds(400,40,50,50);
-        l7.setBounds(450,40,50,50);
-        l8.setBounds(500,40,50,50);
-        l9.setBounds(550,40,50,50);
-        l10.setBounds(600,40,50,50);
-        l11.setBounds(150,90,50,50);
-        l12.setBounds(200,90,50,50);
-        l13.setBounds(250,90,50,50);
-        l14.setBounds(300,90,50,50);
-        l15.setBounds(350,90,50,50);
-        l16.setBounds(400,90,50,50);
-        l17.setBounds(450,90,50,50);
-        l18.setBounds(500,90,50,50);
-        l19.setBounds(550,90,50,50);
-        l20.setBounds(600,90,50,50);
-        l21.setBounds(150,140,50,50);
-        l22.setBounds(200,140,50,50);
-        l23.setBounds(250,140,50,50);
-        l24.setBounds(300,140,50,50);
-        l25.setBounds(350,140,50,50);
-        l26.setBounds(400,140,50,50);
-        l27.setBounds(450,140,50,50);
-        l28.setBounds(500,140,50,50);
-        l29.setBounds(550,140,50,50);
-        l30.setBounds(600,140,50,50);
-        l31.setBounds(150,190,50,50);
-        l32.setBounds(200,190,50,50);
-        l33.setBounds(250,190,50,50);
-        l34.setBounds(300,190,50,50);
-        l35.setBounds(350,190,50,50);
-        l36.setBounds(400,190,50,50);
-        l37.setBounds(450,190,50,50);
-        l38.setBounds(500,190,50,50);
-        l39.setBounds(550,190,50,50);
-        l40.setBounds(600,190,50,50);
-        l41.setBounds(150,240,50,50);
-        l42.setBounds(200,240,50,50);
-        l43.setBounds(250,240,50,50);
-        l44.setBounds(300,240,50,50);
-        l45.setBounds(350,240,50,50);
-        l46.setBounds(400,240,50,50);
-        l47.setBounds(450,240,50,50);
-        l48.setBounds(500,240,50,50);
-        l49.setBounds(550,240,50,50);
-        l50.setBounds(600,240,50,50);
-        l51.setBounds(150,290,50,50);
-        l52.setBounds(200,290,50,50);
-        l53.setBounds(250,290,50,50);
-        l54.setBounds(300,290,50,50);
-        l55.setBounds(350,290,50,50);
-        l56.setBounds(400,290,50,50);
-        l57.setBounds(450,290,50,50);
-        l58.setBounds(500,290,50,50);
-        l59.setBounds(550,290,50,50);
-        l60.setBounds(600,290,50,50);
-        l61.setBounds(150,340,50,50);
-        l62.setBounds(200,340,50,50);
-        l63.setBounds(250,340,50,50);
-        l64.setBounds(300,340,50,50);
-        l65.setBounds(350,340,50,50);
-        l66.setBounds(400,340,50,50);
-        l67.setBounds(450,340,50,50);
-        l68.setBounds(500,340,50,50);
-        l69.setBounds(550,340,50,50);
-        l70.setBounds(600,340,50,50);
-        l71.setBounds(150,390,50,50);
-        l72.setBounds(200,390,50,50);
-        l73.setBounds(250,390,50,50);
-        l74.setBounds(300,390,50,50);
-        l75.setBounds(350,390,50,50);
-        l76.setBounds(400,390,50,50);
-        l77.setBounds(450,390,50,50);
-        l78.setBounds(500,390,50,50);
-        l79.setBounds(550,390,50,50);
-        l80.setBounds(600,390,50,50);
-        l81.setBounds(150,440,50,50);
-        l82.setBounds(200,440,50,50);
-        l83.setBounds(250,440,50,50);
-        l84.setBounds(300,440,50,50);
-        l85.setBounds(350,440,50,50);
-        l86.setBounds(400,440,50,50);
-        l87.setBounds(450,440,50,50);
-        l88.setBounds(500,440,50,50);
-        l89.setBounds(550,440,50,50);
-        l90.setBounds(600,440,50,50);
-        l91.setBounds(150,490,50,50);
-        l92.setBounds(200,490,50,50);
-        l93.setBounds(250,490,50,50);
-        l94.setBounds(300,490,50,50);
-        l95.setBounds(350,490,50,50);
-        l96.setBounds(400,490,50,50);
-        l97.setBounds(450,490,50,50);
-        l98.setBounds(500,490,50,50);
-        l99.setBounds(550,490,50,50);
-        l100.setBounds(600,490,50,50);
-        tp_air.setBounds(45, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        p_air.setBounds(45, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tp_bat.setBounds(185, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        p_bat.setBounds(185, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tp_sub.setBounds(325, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        p_sub.setBounds(325, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tp_cru.setBounds(465, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        p_cru.setBounds(465, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tp_des.setBounds(605, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        p_des.setBounds(605, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tc_air.setBounds(45, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        c_air.setBounds(45, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tc_bat.setBounds(185, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        c_bat.setBounds(185, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tc_sub.setBounds(325, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        c_sub.setBounds(325, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tc_cru.setBounds(465, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        c_cru.setBounds(465, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        tc_des.setBounds(605, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
-        c_des.setBounds(605, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
-        l1.setBorderPainted(false);
-        l2.setBorderPainted(false);
-        l3.setBorderPainted(false);
-        l4.setBorderPainted(false);
-        l5.setBorderPainted(false);
-        l6.setBorderPainted(false);
-        l7.setBorderPainted(false);
-        l8.setBorderPainted(false);
-        l9.setBorderPainted(false);
-        l10.setBorderPainted(false);
-        l11.setBorderPainted(false);
-        l12.setBorderPainted(false);
-        l13.setBorderPainted(false);
-        l14.setBorderPainted(false);
-        l15.setBorderPainted(false);
-        l16.setBorderPainted(false);
-        l17.setBorderPainted(false);
-        l18.setBorderPainted(false);
-        l19.setBorderPainted(false);
-        l20.setBorderPainted(false);
-        l21.setBorderPainted(false);
-        l22.setBorderPainted(false);
-        l23.setBorderPainted(false);
-        l24.setBorderPainted(false);
-        l25.setBorderPainted(false);
-        l26.setBorderPainted(false);
-        l27.setBorderPainted(false);
-        l28.setBorderPainted(false);
-        l29.setBorderPainted(false);
-        l30.setBorderPainted(false);
-        l31.setBorderPainted(false);
-        l32.setBorderPainted(false);
-        l33.setBorderPainted(false);
-        l34.setBorderPainted(false);
-        l35.setBorderPainted(false);
-        l36.setBorderPainted(false);
-        l37.setBorderPainted(false);
-        l38.setBorderPainted(false);
-        l39.setBorderPainted(false);
-        l40.setBorderPainted(false);
-        l41.setBorderPainted(false);
-        l42.setBorderPainted(false);
-        l43.setBorderPainted(false);
-        l44.setBorderPainted(false);
-        l45.setBorderPainted(false);
-        l46.setBorderPainted(false);
-        l47.setBorderPainted(false);
-        l48.setBorderPainted(false);
-        l49.setBorderPainted(false);
-        l50.setBorderPainted(false);
-        l51.setBorderPainted(false);
-        l52.setBorderPainted(false);
-        l53.setBorderPainted(false);
-        l54.setBorderPainted(false);
-        l55.setBorderPainted(false);
-        l56.setBorderPainted(false);
-        l57.setBorderPainted(false);
-        l58.setBorderPainted(false);
-        l59.setBorderPainted(false);
-        l60.setBorderPainted(false);
-        l61.setBorderPainted(false);
-        l62.setBorderPainted(false);
-        l63.setBorderPainted(false);
-        l64.setBorderPainted(false);
-        l65.setBorderPainted(false);
-        l66.setBorderPainted(false);
-        l67.setBorderPainted(false);
-        l68.setBorderPainted(false);
-        l69.setBorderPainted(false);
-        l70.setBorderPainted(false);
-        l71.setBorderPainted(false);
-        l72.setBorderPainted(false);
-        l73.setBorderPainted(false);
-        l74.setBorderPainted(false);
-        l75.setBorderPainted(false);
-        l76.setBorderPainted(false);
-        l77.setBorderPainted(false);
-        l78.setBorderPainted(false);
-        l79.setBorderPainted(false);
-        l80.setBorderPainted(false);
-        l81.setBorderPainted(false);
-        l82.setBorderPainted(false);
-        l83.setBorderPainted(false);
-        l84.setBorderPainted(false);
-        l85.setBorderPainted(false);
-        l86.setBorderPainted(false);
-        l87.setBorderPainted(false);
-        l88.setBorderPainted(false);
-        l89.setBorderPainted(false);
-        l90.setBorderPainted(false);
-        l91.setBorderPainted(false);
-        l92.setBorderPainted(false);
-        l93.setBorderPainted(false);
-        l94.setBorderPainted(false);
-        l95.setBorderPainted(false);
-        l96.setBorderPainted(false);
-        l97.setBorderPainted(false);
-        l98.setBorderPainted(false);
-        l99.setBorderPainted(false);
-        l100.setBorderPainted(false);
-        l1.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l2.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l3.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l4.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l5.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l6.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l7.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l8.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l9.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l10.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l11.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l12.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l13.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l14.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l15.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l16.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l17.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l18.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l19.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l20.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l21.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l22.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l23.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l24.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l25.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l26.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l27.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l28.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l29.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l30.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l31.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l32.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l33.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l34.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l35.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l36.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l37.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l38.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l39.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l40.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l41.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l42.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l43.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l44.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l45.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l46.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l47.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l48.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l49.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l50.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l51.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l52.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l53.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l54.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l55.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l56.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l57.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l58.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l59.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l60.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l61.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l62.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l63.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l64.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l65.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l66.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l67.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l68.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l69.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l70.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l71.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l72.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l73.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l74.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l75.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l76.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l77.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l78.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l79.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l80.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l81.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l82.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l83.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l84.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l85.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l86.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l87.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l88.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l89.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l90.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l91.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l92.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l93.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l94.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l95.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l96.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l97.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l98.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l99.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        l100.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
-        this.play_hit.add(l1);
-        this.play_hit.add(l2);
-        this.play_hit.add(l3);
-        this.play_hit.add(l4);
-        this.play_hit.add(l5);
-        this.play_hit.add(l6);
-        this.play_hit.add(l7);
-        this.play_hit.add(l8);
-        this.play_hit.add(l9);
-        this.play_hit.add(l10);
-        this.play_hit.add(l11);
-        this.play_hit.add(l12);
-        this.play_hit.add(l13);
-        this.play_hit.add(l14);
-        this.play_hit.add(l15);
-        this.play_hit.add(l16);
-        this.play_hit.add(l17);
-        this.play_hit.add(l18);
-        this.play_hit.add(l19);
-        this.play_hit.add(l20);
-        this.play_hit.add(l21);
-        this.play_hit.add(l22);
-        this.play_hit.add(l23);
-        this.play_hit.add(l24);
-        this.play_hit.add(l25);
-        this.play_hit.add(l26);
-        this.play_hit.add(l27);
-        this.play_hit.add(l28);
-        this.play_hit.add(l29);
-        this.play_hit.add(l30);
-        this.play_hit.add(l31);
-        this.play_hit.add(l32);
-        this.play_hit.add(l33);
-        this.play_hit.add(l34);
-        this.play_hit.add(l35);
-        this.play_hit.add(l36);
-        this.play_hit.add(l37);
-        this.play_hit.add(l38);
-        this.play_hit.add(l39);
-        this.play_hit.add(l40);
-        this.play_hit.add(l41);
-        this.play_hit.add(l42);
-        this.play_hit.add(l43);
-        this.play_hit.add(l44);
-        this.play_hit.add(l45);
-        this.play_hit.add(l46);
-        this.play_hit.add(l47);
-        this.play_hit.add(l48);
-        this.play_hit.add(l49);
-        this.play_hit.add(l50);
-        this.play_hit.add(l51);
-        this.play_hit.add(l52);
-        this.play_hit.add(l53);
-        this.play_hit.add(l54);
-        this.play_hit.add(l55);
-        this.play_hit.add(l56);
-        this.play_hit.add(l57);
-        this.play_hit.add(l58);
-        this.play_hit.add(l59);
-        this.play_hit.add(l60);
-        this.play_hit.add(l61);
-        this.play_hit.add(l62);
-        this.play_hit.add(l63);
-        this.play_hit.add(l64);
-        this.play_hit.add(l65);
-        this.play_hit.add(l66);
-        this.play_hit.add(l67);
-        this.play_hit.add(l68);
-        this.play_hit.add(l69);
-        this.play_hit.add(l70);
-        this.play_hit.add(l71);
-        this.play_hit.add(l72);
-        this.play_hit.add(l73);
-        this.play_hit.add(l74);
-        this.play_hit.add(l75);
-        this.play_hit.add(l76);
-        this.play_hit.add(l77);
-        this.play_hit.add(l78);
-        this.play_hit.add(l79);
-        this.play_hit.add(l80);
-        this.play_hit.add(l81);
-        this.play_hit.add(l82);
-        this.play_hit.add(l83);
-        this.play_hit.add(l84);
-        this.play_hit.add(l85);
-        this.play_hit.add(l86);
-        this.play_hit.add(l87);
-        this.play_hit.add(l88);
-        this.play_hit.add(l89);
-        this.play_hit.add(l90);
-        this.play_hit.add(l91);
-        this.play_hit.add(l92);
-        this.play_hit.add(l93);
-        this.play_hit.add(l94);
-        this.play_hit.add(l95);
-        this.play_hit.add(l96);
-        this.play_hit.add(l97);
-        this.play_hit.add(l98);
-        this.play_hit.add(l99);
-        this.play_hit.add(l100);
-        this.play_hit.add(tp_air);
-        this.play_hit.add(p_air);
-        this.play_hit.add(tp_bat);
-        this.play_hit.add(p_bat);
-        this.play_hit.add(tp_sub);
-        this.play_hit.add(p_sub);
-        this.play_hit.add(tp_cru);
-        this.play_hit.add(p_cru);
-        this.play_hit.add(tp_des);
-        this.play_hit.add(p_des);
-        this.j.add(tc_air);
-        this.j.add(c_air);
-        this.j.add(tc_bat);
-        this.j.add(c_bat);
-        this.j.add(tc_sub);
-        this.j.add(c_sub);
-        this.j.add(tc_cru);
-        this.j.add(c_cru);
-        this.j.add(tc_des);
-        this.j.add(c_des);
-        SwingWorker s1 = new SwingWorker()
-        {
-            @Override
-            protected String doInBackground() throws Exception
-            {
-                try
-                {
-                    FileWriter fw = new FileWriter("logs/logfile.txt", true);
-                    fw.write("Aircraft Carrier : \n");
-                    for(int i = 0 ; i < 5 ; i ++)
-                    {
-                        fw.write(aircraft_pos[i] + "\n");
-                    }
-                    fw.write("Battleship : \n");
-                    for(int i = 0 ; i < 4 ; i ++)
-                    {
-                        fw.write(battleship_pos[i] + "\n");
-                    }
-                    fw.write("Submarine : \n");
-                    for(int i = 0 ; i < 3 ; i ++)
-                    {
-                        fw.write(submarine_pos[i] + "\n");
-                    }
-                    fw.write("Cruiser : \n");
-                    for(int i = 0 ; i < 3 ; i ++)
-                    {
-                        fw.write(cruiser_pos[i] + "\n");
-                    }
-                    fw.write("Destroyer : \n");
-                    for(int i = 0 ; i < 2 ; i ++)
-                    {
-                        fw.write(destroyer_pos[i] + "\n");
-                    }
-                    fw.close();
-                }
-                catch(Exception ex)
-                {
-                    JOptionPane.showMessageDialog(j, ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
-                }
-                com_gen();
-                while (true)
-                {
-                    SwingUtilities.invokeLater(new Runnable(){
-                        public void run()
-                        {
-                            PlayArea.repaint();
-                            j.repaint();
-                            loading.repaint();
-                        }
-                    });
-                    r.setSeed(seed);
-                    int success = 0;
-                    int or = Math.abs(r.nextInt());
-                    or = or % 2;
-                    if(or == 0)
-                    {
-                        int temp_flag = 0;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 60;
-                        int pos1 = com_air_place_h[pos];
-                        int pos2 = pos1 + 1;
-                        int pos3 = pos2 + 1;
-                        int pos4 = pos3 + 1;
-                        int pos5 = pos4 + 1;
-                        com_aircraft_pos[0] = pos1;
-                        com_aircraft_pos[1] = pos2;
-                        com_aircraft_pos[2] = pos3;
-                        com_aircraft_pos[3] = pos4;
-                        com_aircraft_pos[4] = pos5;
-                        success = 1;
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                    else if (or == 1)
-                    {
-                        int temp_flag = 0;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 60;
-                        int pos1 = com_air_place_v[pos];
-                        int pos2 = pos1 + 10;
-                        int pos3 = pos2 + 10;
-                        int pos4 = pos3 + 10;
-                        int pos5 = pos4 + 10;
-                        com_aircraft_pos[0] = pos1;
-                        com_aircraft_pos[1] = pos2;
-                        com_aircraft_pos[2] = pos3;
-                        com_aircraft_pos[3] = pos4;
-                        com_aircraft_pos[4] = pos5;
-                        success = 1;
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                }
-                j.revalidate();
-                j.repaint();
-                int progress = 1;
-                while(progress <= 20)
-                {
-                    try
-                    {
-                        pb.setValue(progress);
-                        j.revalidate();
-                        j.repaint();
-                        Thread.sleep(50);
-                        progress++;
-                    }
-                    catch (Exception ex)
-                    {
 
-                    }
-                }
-                while (true)
-                {
-                    int p_flag = 0;
-                    int success = 0;
-                    int or = Math.abs(r.nextInt());
-                    or = or % 2;
-                    if(or == 0)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 70;
-                        int pos1 = com_bat_place_h[pos];
-                        int pos2 = pos1 + 1;
-                        int pos3 = pos2 + 1;
-                        int pos4 = pos3 + 1;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-                            
-                            
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2 && com_aircraft_pos[i] != pos3 && com_aircraft_pos[i] != pos4)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 6)
-                        {
-                            com_battleship_pos[0] = pos1;
-                            com_battleship_pos[1] = pos2;
-                            com_battleship_pos[2] = pos3;
-                            com_battleship_pos[3] = pos4;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                    else if (or == 1)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 70;
-                        int pos1 = com_bat_place_v[pos];
-                        int pos2 = pos1 + 10;
-                        int pos3 = pos2 + 10;
-                        int pos4 = pos3 + 10;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-                            
-                            
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2 && com_aircraft_pos[i] != pos3 && com_aircraft_pos[i] != pos4)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 6)
-                        {
-                            com_battleship_pos[0] = pos1;
-                            com_battleship_pos[1] = pos2;
-                            com_battleship_pos[2] = pos3;
-                            com_battleship_pos[3] = pos4;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                }
-                while(progress <= 40)
-                {
-                    try
-                    {
-                        pb.setValue(progress);
-                        j.revalidate();
-                        j.repaint();
-                        Thread.sleep(50);
-                        progress++;
-                    }
-                    catch (Exception ex)
-                    {
-
-                    }
-                }
-                while (true)
-                {
-                    int p_flag = 0;
-                    int success = 0;
-                    int or = Math.abs(r.nextInt());
-                    or = or % 2;
-                    if(or == 0)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 80;
-                        int pos1 = com_sub_place_h[pos];
-                        int pos2 = pos1 + 1;
-                        int pos3 = pos2 + 1;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-                            
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2 && com_aircraft_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 3; i ++)
-                        {
-                            
-                            if(com_battleship_pos[i] != pos1 && com_battleship_pos[i] != pos2 && com_battleship_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 10)
-                        {
-                            com_submarine_pos[0] = pos1;
-                            com_submarine_pos[1] = pos2;
-                            com_submarine_pos[2] = pos3;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                    else if (or == 1)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 80;
-                        int pos1 = com_sub_place_v[pos];
-                        int pos2 = pos1 + 10;
-                        int pos3 = pos2 + 10;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-                            
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2 && com_aircraft_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 3; i ++)
-                        {
-                            
-                            if(com_battleship_pos[i] != pos1 && com_battleship_pos[i] != pos2 && com_battleship_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 10)
-                        {
-                            com_submarine_pos[0] = pos1;
-                            com_submarine_pos[1] = pos2;
-                            com_submarine_pos[2] = pos3;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                }
-                while(progress <= 60)
-                {
-                    try
-                    {
-                        pb.setValue(progress);
-                        j.revalidate();
-                        j.repaint();
-                        Thread.sleep(50);
-                        progress++;
-                    }
-                    catch (Exception ex)
-                    {
-
-                    }
-                }
-                while (true)
-                {
-                    int p_flag = 0;
-                    int success = 0;
-                    int or = Math.abs(r.nextInt());
-                    or = or % 2;
-                    if(or == 0)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 80;
-                        int pos1 = com_cru_place_h[pos];
-                        int pos2 = pos1 + 1;
-                        int pos3 = pos2 + 1;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-                            
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2 && com_aircraft_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 3; i ++)
-                        {
-                            
-                            if(com_battleship_pos[i] != pos1 && com_battleship_pos[i] != pos2 && com_battleship_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 2; i ++)
-                        {
-                            
-                            if(com_submarine_pos[i] != pos1 && com_submarine_pos[i] != pos2 && com_submarine_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 13)
-                        {
-                            com_cruiser_pos[0] = pos1;
-                            com_cruiser_pos[1] = pos2;
-                            com_cruiser_pos[2] = pos3;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                    else if (or == 1)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 80;
-                        int pos1 = com_cru_place_v[pos];
-                        int pos2 = pos1 + 10;
-                        int pos3 = pos2 + 10;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-                            
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2 && com_aircraft_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 3; i ++)
-                        {
-                            
-                            if(com_battleship_pos[i] != pos1 && com_battleship_pos[i] != pos2 && com_battleship_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 2; i ++)
-                        {
-                            
-                            if(com_submarine_pos[i] != pos1 && com_submarine_pos[i] != pos2 && com_submarine_pos[i] != pos3)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 13)
-                        {
-                            com_cruiser_pos[0] = pos1;
-                            com_cruiser_pos[1] = pos2;
-                            com_cruiser_pos[2] = pos3;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                }
-                while(progress <= 80)
-                {
-                    try
-                    {
-                        pb.setValue(progress);
-                        j.revalidate();
-                        j.repaint();
-                        Thread.sleep(50);
-                        progress++;
-                    }
-                    catch (Exception ex)
-                    {
-
-                    }
-                }
-                while (true)
-                {
-                    int p_flag = 0;
-                    int success = 0;
-                    int or = Math.abs(r.nextInt());
-                    or = or % 2;
-                    if(or == 0)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 90;
-                        int pos1 = com_des_place_h[pos];
-                        int pos2 = pos1 + 1;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 3; i ++)
-                        {
-
-                            if(com_battleship_pos[i] != pos1 && com_battleship_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 2; i ++)
-                        {
-
-                            if(com_submarine_pos[i] != pos1 && com_submarine_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 2; i ++)
-                        {
-
-                            if(com_cruiser_pos[i] != pos1 && com_cruiser_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 16)
-                        {
-                            com_destroyer_pos[0] = pos1;
-                            com_destroyer_pos[1] = pos2;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                    else if (or == 1)
-                    {
-                        int temp_flag = 1;
-                        int pos = Math.abs(r.nextInt());
-                        pos = pos % 90;
-                        int pos1 = com_des_place_v[pos];
-                        int pos2 = pos1 + 10;
-                        for(int i = 0; i <= 4; i ++)
-                        {
-
-                            if(com_aircraft_pos[i] != pos1 && com_aircraft_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 3; i ++)
-                        {
-
-                            if(com_battleship_pos[i] != pos1 && com_battleship_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 2; i ++)
-                        {
-
-                            if(com_submarine_pos[i] != pos1 && com_submarine_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        for(int i = 0; i <= 2; i ++)
-                        {
-
-                            if(com_cruiser_pos[i] != pos1 && com_cruiser_pos[i] != pos2)
-                            {
-                                temp_flag++;
-                            }
-                        }
-                        System.out.println(temp_flag);
-                        if(temp_flag == 16)
-                        {
-                            com_destroyer_pos[0] = pos1;
-                            com_destroyer_pos[1] = pos2;
-                            success = 1;
-                        }
-                        if (success == 1)
-                        {
-                            break;
-                        }
-                    }
-                }
-                while(progress <= 100)
-                {
-                    try
-                    {
-                        pb.setValue(progress);
-                        j.revalidate();
-                        j.repaint();
-                        Thread.sleep(50);
-                        progress++;
-                    }
-                    catch (Exception ex)
-                    {
-
-                    }
-                }
-//                System.out.println("Computer Positions:\nAircraft Carrier:");
-//                System.out.println(com_aircraft_pos[0]);
-//                System.out.println(com_aircraft_pos[1]);
-//                System.out.println(com_aircraft_pos[2]);
-//                System.out.println(com_aircraft_pos[3]);
-//                System.out.println(com_aircraft_pos[4]);
-//                System.out.println("Battleship:");
-//                System.out.println(com_battleship_pos[0]);
-//                System.out.println(com_battleship_pos[1]);
-//                System.out.println(com_battleship_pos[2]);
-//                System.out.println(com_battleship_pos[3]);
-//                System.out.println("Submarine:");
-//                System.out.println(com_submarine_pos[0]);
-//                System.out.println(com_submarine_pos[1]);
-//                System.out.println(com_submarine_pos[2]);
-//                System.out.println("Cruiser:");
-//                System.out.println(com_cruiser_pos[0]);
-//                System.out.println(com_cruiser_pos[1]);
-//                System.out.println(com_cruiser_pos[2]);
-//                System.out.println("Destroyer:");
-//                System.out.println(com_destroyer_pos[0]);
-//                System.out.println(com_destroyer_pos[1]);
-                load.setText("Switching between bases");
-                return null;
-            }
-            @Override
-            protected void done()
-            {
-                who_is_playing();
-            }
-        };
-        s1.execute();
-    }
-    void com_gen()
-    {
-        for(int i = 0, j = 1 ; i < 60; i++)
-        {
-            com_air_place_h[i] = j;
-            System.out.println(com_air_place_h[i]);
-            if (j == 6)
-            {
-                j = 11;
-            }
-            else if (j == 16)
-            {
-                j = 21;
-            }
-            else if (j == 26)
-            {
-                j = 31;
-            }
-            else if (j == 36)
-            {
-                j = 41;
-            }
-            else if (j == 46)
-            {
-                j = 51;
-            }
-            else if (j == 56)
-            {
-                j = 61;
-            }
-            else if (j == 66)
-            {
-                j = 71;
-            }
-            else if (j == 76)
-            {
-                j = 81;
-            }
-            else if (j == 86)
-            {
-                j = 91;
-            }
-            else if (j == 96)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 60; i++)
-        {
-            com_air_place_v[i] = j;
-            System.out.println(com_air_place_v[i]);
-            if (j == 60)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 70; i++)
-        {
-            com_bat_place_h[i] = j;
-            System.out.println(com_bat_place_h[i]);
-            if (j == 7)
-            {
-                j = 11;
-            }
-            else if (j == 17)
-            {
-                j = 21;
-            }
-            else if (j == 27)
-            {
-                j = 31;
-            }
-            else if (j == 37)
-            {
-                j = 41;
-            }
-            else if (j == 47)
-            {
-                j = 51;
-            }
-            else if (j == 57)
-            {
-                j = 61;
-            }
-            else if (j == 67)
-            {
-                j = 71;
-            }
-            else if (j == 77)
-            {
-                j = 81;
-            }
-            else if (j == 87)
-            {
-                j = 91;
-            }
-            else if (j == 97)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 70; i++)
-        {
-            com_bat_place_v[i] = j;
-            System.out.println(com_bat_place_v[i]);
-            if (j == 70)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 80; i++)
-        {
-            com_sub_place_h[i] = j;
-            System.out.println(com_sub_place_h[i]);
-            if (j == 8)
-            {
-                j = 11;
-            }
-            else if (j == 18)
-            {
-                j = 21;
-            }
-            else if (j == 28)
-            {
-                j = 31;
-            }
-            else if (j == 38)
-            {
-                j = 41;
-            }
-            else if (j == 48)
-            {
-                j = 51;
-            }
-            else if (j == 58)
-            {
-                j = 61;
-            }
-            else if (j == 68)
-            {
-                j = 71;
-            }
-            else if (j == 78)
-            {
-                j = 81;
-            }
-            else if (j == 88)
-            {
-                j = 91;
-            }
-            else if (j == 98)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 80; i++)
-        {
-            com_sub_place_v[i] = j;
-            System.out.println(com_sub_place_v[i]);
-            if (j == 80)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 80; i++)
-        {
-            com_cru_place_h[i] = j;
-            System.out.println(com_cru_place_h[i]);
-            if (j == 8)
-            {
-                j = 11;
-            }
-            else if (j == 18)
-            {
-                j = 21;
-            }
-            else if (j == 28)
-            {
-                j = 31;
-            }
-            else if (j == 38)
-            {
-                j = 41;
-            }
-            else if (j == 48)
-            {
-                j = 51;
-            }
-            else if (j == 58)
-            {
-                j = 61;
-            }
-            else if (j == 68)
-            {
-                j = 71;
-            }
-            else if (j == 78)
-            {
-                j = 81;
-            }
-            else if (j == 88)
-            {
-                j = 91;
-            }
-            else if (j == 98)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 80; i++)
-        {
-            com_cru_place_v[i] = j;
-            System.out.println(com_cru_place_v[i]);
-            if (j == 80)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 90; i++)
-        {
-            com_des_place_h[i] = j;
-            System.out.println(com_des_place_h[i]);
-            if (j == 9)
-            {
-                j = 11;
-            }
-            else if (j == 19)
-            {
-                j = 21;
-            }
-            else if (j == 29)
-            {
-                j = 31;
-            }
-            else if (j == 39)
-            {
-                j = 41;
-            }
-            else if (j == 49)
-            {
-                j = 51;
-            }
-            else if (j == 59)
-            {
-                j = 61;
-            }
-            else if (j == 69)
-            {
-                j = 71;
-            }
-            else if (j == 79)
-            {
-                j = 81;
-            }
-            else if (j == 89)
-            {
-                j = 91;
-            }
-            else if (j == 99)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-        for(int i = 0, j = 1 ; i < 90; i++)
-        {
-            com_des_place_v[i] = j;
-            System.out.println(com_des_place_v[i]);
-            if (j == 90)
-            {
-                break;
-            }
-            else
-            {
-                j++;
-            }
-        }
-    }
     public static void main(String[] args)
     {
         Main m = new Main();
