+++ src/generation.java
@@ -1,0 +1,670 @@
+import javax.swing.*;
+import java.awt.*;
+
+public class generation extends data
+{
+    public void generate()
+    {
+        j.remove(tip);
+        convert_player();
+        Icon i = new ImageIcon("water.png");
+        l1 = new JButton(i);
+        l2 = new JButton(i);
+        l3 = new JButton(i);
+        l4 = new JButton(i);
+        l5 = new JButton(i);
+        l6 = new JButton(i);
+        l7 = new JButton(i);
+        l8 = new JButton(i);
+        l9 = new JButton(i);
+        l10 = new JButton(i);
+        l11 = new JButton(i);
+        l12 = new JButton(i);
+        l13 = new JButton(i);
+        l14 = new JButton(i);
+        l15 = new JButton(i);
+        l16 = new JButton(i);
+        l17 = new JButton(i);
+        l18 = new JButton(i);
+        l19 = new JButton(i);
+        l20 = new JButton(i);
+        l21 = new JButton(i);
+        l22 = new JButton(i);
+        l23 = new JButton(i);
+        l24 = new JButton(i);
+        l25 = new JButton(i);
+        l26 = new JButton(i);
+        l27 = new JButton(i);
+        l28 = new JButton(i);
+        l29 = new JButton(i);
+        l30 = new JButton(i);
+        l31 = new JButton(i);
+        l32 = new JButton(i);
+        l33 = new JButton(i);
+        l34 = new JButton(i);
+        l35 = new JButton(i);
+        l36 = new JButton(i);
+        l37 = new JButton(i);
+        l38 = new JButton(i);
+        l39 = new JButton(i);
+        l40 = new JButton(i);
+        l41 = new JButton(i);
+        l42 = new JButton(i);
+        l43 = new JButton(i);
+        l44 = new JButton(i);
+        l45 = new JButton(i);
+        l46 = new JButton(i);
+        l47 = new JButton(i);
+        l48 = new JButton(i);
+        l49 = new JButton(i);
+        l50 = new JButton(i);
+        l51 = new JButton(i);
+        l52 = new JButton(i);
+        l53 = new JButton(i);
+        l54 = new JButton(i);
+        l55 = new JButton(i);
+        l56 = new JButton(i);
+        l57 = new JButton(i);
+        l58 = new JButton(i);
+        l59 = new JButton(i);
+        l60 = new JButton(i);
+        l61 = new JButton(i);
+        l62 = new JButton(i);
+        l63 = new JButton(i);
+        l64 = new JButton(i);
+        l65 = new JButton(i);
+        l66 = new JButton(i);
+        l67 = new JButton(i);
+        l68 = new JButton(i);
+        l69 = new JButton(i);
+        l70 = new JButton(i);
+        l71 = new JButton(i);
+        l72 = new JButton(i);
+        l73 = new JButton(i);
+        l74 = new JButton(i);
+        l75 = new JButton(i);
+        l76 = new JButton(i);
+        l77 = new JButton(i);
+        l78 = new JButton(i);
+        l79 = new JButton(i);
+        l80 = new JButton(i);
+        l81 = new JButton(i);
+        l82 = new JButton(i);
+        l83 = new JButton(i);
+        l84 = new JButton(i);
+        l85 = new JButton(i);
+        l86 = new JButton(i);
+        l87 = new JButton(i);
+        l88 = new JButton(i);
+        l89 = new JButton(i);
+        l90 = new JButton(i);
+        l91 = new JButton(i);
+        l92 = new JButton(i);
+        l93 = new JButton(i);
+        l94 = new JButton(i);
+        l95 = new JButton(i);
+        l96 = new JButton(i);
+        l97 = new JButton(i);
+        l98 = new JButton(i);
+        l99 = new JButton(i);
+        l100 = new JButton(i);
+        tp_air = new JLabel("Aircraft Carrier");
+        tp_bat = new JLabel("Battleship");
+        tp_sub = new JLabel("Submarine");
+        tp_cru = new JLabel("Cruiser");
+        tp_des = new JLabel("Destroyer");
+        tp_air.setForeground(Color.WHITE);
+        tp_bat.setForeground(Color.WHITE);
+        tp_sub.setForeground(Color.WHITE);
+        tp_cru.setForeground(Color.WHITE);
+        tp_des.setForeground(Color.WHITE);
+        p_air = new JLabel(safe);
+        p_bat = new JLabel(safe);
+        p_sub = new JLabel(safe);
+        p_cru = new JLabel(safe);
+        p_des = new JLabel(safe);
+        tc_air = new JLabel("Aircraft Carrier");
+        tc_bat = new JLabel("Battleship");
+        tc_sub = new JLabel("Submarine");
+        tc_cru = new JLabel("Cruiser");
+        tc_des = new JLabel("Destroyer");
+        tc_air.setForeground(Color.WHITE);
+        tc_bat.setForeground(Color.WHITE);
+        tc_sub.setForeground(Color.WHITE);
+        tc_cru.setForeground(Color.WHITE);
+        tc_des.setForeground(Color.WHITE);
+        c_air = new JLabel(hit);
+        c_bat = new JLabel(hit);
+        c_sub = new JLabel(hit);
+        c_cru = new JLabel(hit);
+        c_des = new JLabel(hit);
+        l1.setBounds(150,40,50,50);
+        l2.setBounds(200,40,50,50);
+        l3.setBounds(250,40,50,50);
+        l4.setBounds(300,40,50,50);
+        l5.setBounds(350,40,50,50);
+        l6.setBounds(400,40,50,50);
+        l7.setBounds(450,40,50,50);
+        l8.setBounds(500,40,50,50);
+        l9.setBounds(550,40,50,50);
+        l10.setBounds(600,40,50,50);
+        l11.setBounds(150,90,50,50);
+        l12.setBounds(200,90,50,50);
+        l13.setBounds(250,90,50,50);
+        l14.setBounds(300,90,50,50);
+        l15.setBounds(350,90,50,50);
+        l16.setBounds(400,90,50,50);
+        l17.setBounds(450,90,50,50);
+        l18.setBounds(500,90,50,50);
+        l19.setBounds(550,90,50,50);
+        l20.setBounds(600,90,50,50);
+        l21.setBounds(150,140,50,50);
+        l22.setBounds(200,140,50,50);
+        l23.setBounds(250,140,50,50);
+        l24.setBounds(300,140,50,50);
+        l25.setBounds(350,140,50,50);
+        l26.setBounds(400,140,50,50);
+        l27.setBounds(450,140,50,50);
+        l28.setBounds(500,140,50,50);
+        l29.setBounds(550,140,50,50);
+        l30.setBounds(600,140,50,50);
+        l31.setBounds(150,190,50,50);
+        l32.setBounds(200,190,50,50);
+        l33.setBounds(250,190,50,50);
+        l34.setBounds(300,190,50,50);
+        l35.setBounds(350,190,50,50);
+        l36.setBounds(400,190,50,50);
+        l37.setBounds(450,190,50,50);
+        l38.setBounds(500,190,50,50);
+        l39.setBounds(550,190,50,50);
+        l40.setBounds(600,190,50,50);
+        l41.setBounds(150,240,50,50);
+        l42.setBounds(200,240,50,50);
+        l43.setBounds(250,240,50,50);
+        l44.setBounds(300,240,50,50);
+        l45.setBounds(350,240,50,50);
+        l46.setBounds(400,240,50,50);
+        l47.setBounds(450,240,50,50);
+        l48.setBounds(500,240,50,50);
+        l49.setBounds(550,240,50,50);
+        l50.setBounds(600,240,50,50);
+        l51.setBounds(150,290,50,50);
+        l52.setBounds(200,290,50,50);
+        l53.setBounds(250,290,50,50);
+        l54.setBounds(300,290,50,50);
+        l55.setBounds(350,290,50,50);
+        l56.setBounds(400,290,50,50);
+        l57.setBounds(450,290,50,50);
+        l58.setBounds(500,290,50,50);
+        l59.setBounds(550,290,50,50);
+        l60.setBounds(600,290,50,50);
+        l61.setBounds(150,340,50,50);
+        l62.setBounds(200,340,50,50);
+        l63.setBounds(250,340,50,50);
+        l64.setBounds(300,340,50,50);
+        l65.setBounds(350,340,50,50);
+        l66.setBounds(400,340,50,50);
+        l67.setBounds(450,340,50,50);
+        l68.setBounds(500,340,50,50);
+        l69.setBounds(550,340,50,50);
+        l70.setBounds(600,340,50,50);
+        l71.setBounds(150,390,50,50);
+        l72.setBounds(200,390,50,50);
+        l73.setBounds(250,390,50,50);
+        l74.setBounds(300,390,50,50);
+        l75.setBounds(350,390,50,50);
+        l76.setBounds(400,390,50,50);
+        l77.setBounds(450,390,50,50);
+        l78.setBounds(500,390,50,50);
+        l79.setBounds(550,390,50,50);
+        l80.setBounds(600,390,50,50);
+        l81.setBounds(150,440,50,50);
+        l82.setBounds(200,440,50,50);
+        l83.setBounds(250,440,50,50);
+        l84.setBounds(300,440,50,50);
+        l85.setBounds(350,440,50,50);
+        l86.setBounds(400,440,50,50);
+        l87.setBounds(450,440,50,50);
+        l88.setBounds(500,440,50,50);
+        l89.setBounds(550,440,50,50);
+        l90.setBounds(600,440,50,50);
+        l91.setBounds(150,490,50,50);
+        l92.setBounds(200,490,50,50);
+        l93.setBounds(250,490,50,50);
+        l94.setBounds(300,490,50,50);
+        l95.setBounds(350,490,50,50);
+        l96.setBounds(400,490,50,50);
+        l97.setBounds(450,490,50,50);
+        l98.setBounds(500,490,50,50);
+        l99.setBounds(550,490,50,50);
+        l100.setBounds(600,490,50,50);
+        tp_air.setBounds(45, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        p_air.setBounds(45, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tp_bat.setBounds(185, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        p_bat.setBounds(185, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tp_sub.setBounds(325, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        p_sub.setBounds(325, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tp_cru.setBounds(465, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        p_cru.setBounds(465, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tp_des.setBounds(605, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        p_des.setBounds(605, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tc_air.setBounds(45, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        c_air.setBounds(45, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tc_bat.setBounds(185, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        c_bat.setBounds(185, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tc_sub.setBounds(325, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        c_sub.setBounds(325, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tc_cru.setBounds(465, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        c_cru.setBounds(465, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        tc_des.setBounds(605, 571, tp_air.getPreferredSize().width, tp_air.getPreferredSize().height);
+        c_des.setBounds(605, 591, p_air.getPreferredSize().width, p_air.getPreferredSize().height);
+        l1.setBorderPainted(false);
+        l2.setBorderPainted(false);
+        l3.setBorderPainted(false);
+        l4.setBorderPainted(false);
+        l5.setBorderPainted(false);
+        l6.setBorderPainted(false);
+        l7.setBorderPainted(false);
+        l8.setBorderPainted(false);
+        l9.setBorderPainted(false);
+        l10.setBorderPainted(false);
+        l11.setBorderPainted(false);
+        l12.setBorderPainted(false);
+        l13.setBorderPainted(false);
+        l14.setBorderPainted(false);
+        l15.setBorderPainted(false);
+        l16.setBorderPainted(false);
+        l17.setBorderPainted(false);
+        l18.setBorderPainted(false);
+        l19.setBorderPainted(false);
+        l20.setBorderPainted(false);
+        l21.setBorderPainted(false);
+        l22.setBorderPainted(false);
+        l23.setBorderPainted(false);
+        l24.setBorderPainted(false);
+        l25.setBorderPainted(false);
+        l26.setBorderPainted(false);
+        l27.setBorderPainted(false);
+        l28.setBorderPainted(false);
+        l29.setBorderPainted(false);
+        l30.setBorderPainted(false);
+        l31.setBorderPainted(false);
+        l32.setBorderPainted(false);
+        l33.setBorderPainted(false);
+        l34.setBorderPainted(false);
+        l35.setBorderPainted(false);
+        l36.setBorderPainted(false);
+        l37.setBorderPainted(false);
+        l38.setBorderPainted(false);
+        l39.setBorderPainted(false);
+        l40.setBorderPainted(false);
+        l41.setBorderPainted(false);
+        l42.setBorderPainted(false);
+        l43.setBorderPainted(false);
+        l44.setBorderPainted(false);
+        l45.setBorderPainted(false);
+        l46.setBorderPainted(false);
+        l47.setBorderPainted(false);
+        l48.setBorderPainted(false);
+        l49.setBorderPainted(false);
+        l50.setBorderPainted(false);
+        l51.setBorderPainted(false);
+        l52.setBorderPainted(false);
+        l53.setBorderPainted(false);
+        l54.setBorderPainted(false);
+        l55.setBorderPainted(false);
+        l56.setBorderPainted(false);
+        l57.setBorderPainted(false);
+        l58.setBorderPainted(false);
+        l59.setBorderPainted(false);
+        l60.setBorderPainted(false);
+        l61.setBorderPainted(false);
+        l62.setBorderPainted(false);
+        l63.setBorderPainted(false);
+        l64.setBorderPainted(false);
+        l65.setBorderPainted(false);
+        l66.setBorderPainted(false);
+        l67.setBorderPainted(false);
+        l68.setBorderPainted(false);
+        l69.setBorderPainted(false);
+        l70.setBorderPainted(false);
+        l71.setBorderPainted(false);
+        l72.setBorderPainted(false);
+        l73.setBorderPainted(false);
+        l74.setBorderPainted(false);
+        l75.setBorderPainted(false);
+        l76.setBorderPainted(false);
+        l77.setBorderPainted(false);
+        l78.setBorderPainted(false);
+        l79.setBorderPainted(false);
+        l80.setBorderPainted(false);
+        l81.setBorderPainted(false);
+        l82.setBorderPainted(false);
+        l83.setBorderPainted(false);
+        l84.setBorderPainted(false);
+        l85.setBorderPainted(false);
+        l86.setBorderPainted(false);
+        l87.setBorderPainted(false);
+        l88.setBorderPainted(false);
+        l89.setBorderPainted(false);
+        l90.setBorderPainted(false);
+        l91.setBorderPainted(false);
+        l92.setBorderPainted(false);
+        l93.setBorderPainted(false);
+        l94.setBorderPainted(false);
+        l95.setBorderPainted(false);
+        l96.setBorderPainted(false);
+        l97.setBorderPainted(false);
+        l98.setBorderPainted(false);
+        l99.setBorderPainted(false);
+        l100.setBorderPainted(false);
+        l1.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l2.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l3.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l4.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l5.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l6.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l7.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l8.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l9.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l10.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l11.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l12.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l13.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l14.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l15.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l16.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l17.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l18.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l19.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l20.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l21.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l22.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l23.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l24.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l25.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l26.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l27.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l28.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l29.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l30.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l31.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l32.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l33.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l34.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l35.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l36.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l37.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l38.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l39.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l40.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l41.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l42.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l43.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l44.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l45.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l46.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l47.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l48.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l49.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l50.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l51.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l52.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l53.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l54.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l55.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l56.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l57.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l58.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l59.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l60.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l61.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l62.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l63.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l64.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l65.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l66.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l67.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l68.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l69.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l70.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l71.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l72.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l73.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l74.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l75.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l76.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l77.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l78.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l79.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l80.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l81.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l82.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l83.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l84.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l85.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l86.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l87.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l88.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l89.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l90.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l91.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l92.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l93.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l94.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l95.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l96.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l97.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l98.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l99.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        l100.setBorder(BorderFactory.createLineBorder(Color.BLACK, 1, true));
+        this.play_hit.add(l1);
+        this.play_hit.add(l2);
+        this.play_hit.add(l3);
+        this.play_hit.add(l4);
+        this.play_hit.add(l5);
+        this.play_hit.add(l6);
+        this.play_hit.add(l7);
+        this.play_hit.add(l8);
+        this.play_hit.add(l9);
+        this.play_hit.add(l10);
+        this.play_hit.add(l11);
+        this.play_hit.add(l12);
+        this.play_hit.add(l13);
+        this.play_hit.add(l14);
+        this.play_hit.add(l15);
+        this.play_hit.add(l16);
+        this.play_hit.add(l17);
+        this.play_hit.add(l18);
+        this.play_hit.add(l19);
+        this.play_hit.add(l20);
+        this.play_hit.add(l21);
+        this.play_hit.add(l22);
+        this.play_hit.add(l23);
+        this.play_hit.add(l24);
+        this.play_hit.add(l25);
+        this.play_hit.add(l26);
+        this.play_hit.add(l27);
+        this.play_hit.add(l28);
+        this.play_hit.add(l29);
+        this.play_hit.add(l30);
+        this.play_hit.add(l31);
+        this.play_hit.add(l32);
+        this.play_hit.add(l33);
+        this.play_hit.add(l34);
+        this.play_hit.add(l35);
+        this.play_hit.add(l36);
+        this.play_hit.add(l37);
+        this.play_hit.add(l38);
+        this.play_hit.add(l39);
+        this.play_hit.add(l40);
+        this.play_hit.add(l41);
+        this.play_hit.add(l42);
+        this.play_hit.add(l43);
+        this.play_hit.add(l44);
+        this.play_hit.add(l45);
+        this.play_hit.add(l46);
+        this.play_hit.add(l47);
+        this.play_hit.add(l48);
+        this.play_hit.add(l49);
+        this.play_hit.add(l50);
+        this.play_hit.add(l51);
+        this.play_hit.add(l52);
+        this.play_hit.add(l53);
+        this.play_hit.add(l54);
+        this.play_hit.add(l55);
+        this.play_hit.add(l56);
+        this.play_hit.add(l57);
+        this.play_hit.add(l58);
+        this.play_hit.add(l59);
+        this.play_hit.add(l60);
+        this.play_hit.add(l61);
+        this.play_hit.add(l62);
+        this.play_hit.add(l63);
+        this.play_hit.add(l64);
+        this.play_hit.add(l65);
+        this.play_hit.add(l66);
+        this.play_hit.add(l67);
+        this.play_hit.add(l68);
+        this.play_hit.add(l69);
+        this.play_hit.add(l70);
+        this.play_hit.add(l71);
+        this.play_hit.add(l72);
+        this.play_hit.add(l73);
+        this.play_hit.add(l74);
+        this.play_hit.add(l75);
+        this.play_hit.add(l76);
+        this.play_hit.add(l77);
+        this.play_hit.add(l78);
+        this.play_hit.add(l79);
+        this.play_hit.add(l80);
+        this.play_hit.add(l81);
+        this.play_hit.add(l82);
+        this.play_hit.add(l83);
+        this.play_hit.add(l84);
+        this.play_hit.add(l85);
+        this.play_hit.add(l86);
+        this.play_hit.add(l87);
+        this.play_hit.add(l88);
+        this.play_hit.add(l89);
+        this.play_hit.add(l90);
+        this.play_hit.add(l91);
+        this.play_hit.add(l92);
+        this.play_hit.add(l93);
+        this.play_hit.add(l94);
+        this.play_hit.add(l95);
+        this.play_hit.add(l96);
+        this.play_hit.add(l97);
+        this.play_hit.add(l98);
+        this.play_hit.add(l99);
+        this.play_hit.add(l100);
+        this.play_hit.add(tp_air);
+        this.play_hit.add(p_air);
+        this.play_hit.add(tp_bat);
+        this.play_hit.add(p_bat);
+        this.play_hit.add(tp_sub);
+        this.play_hit.add(p_sub);
+        this.play_hit.add(tp_cru);
+        this.play_hit.add(p_cru);
+        this.play_hit.add(tp_des);
+        this.play_hit.add(p_des);
+        this.j.add(tc_air);
+        this.j.add(c_air);
+        this.j.add(tc_bat);
+        this.j.add(c_bat);
+        this.j.add(tc_sub);
+        this.j.add(c_sub);
+        this.j.add(tc_cru);
+        this.j.add(c_cru);
+        this.j.add(tc_des);
+        this.j.add(c_des);
+    }
+    void convert_player()
+    {
+        for(int i = 0 ; i < 5 ; i ++)
+        {
+            play_aircraft_pos[i] = Integer.parseInt(aircraft_pos[i].substring(1));
+        }
+        for(int i = 0 ; i < 4 ; i ++)
+        {
+            play_battleship_pos[i] = Integer.parseInt(battleship_pos[i].substring(1));
+        }
+        for(int i = 0 ; i < 3 ; i ++)
+        {
+            play_submarine_pos[i] = Integer.parseInt(submarine_pos[i].substring(1));
+        }
+        for(int i = 0 ; i < 3 ; i ++)
+        {
+            play_cruiser_pos[i] = Integer.parseInt(cruiser_pos[i].substring(1));
+        }
+        for(int i = 0 ; i < 2 ; i ++)
+        {
+            play_destroyer_pos[i] = Integer.parseInt(destroyer_pos[i].substring(1));
+        }
+        if((play_aircraft_pos[1] - play_aircraft_pos[0]) == 1)
+        {
+            air_or = 0;
+        }
+        else
+        {
+            air_or = 1;
+        }
+        if((play_battleship_pos[1] - play_battleship_pos[0]) == 1)
+        {
+            bat_or = 0;
+        }
+        else
+        {
+            bat_or = 1;
+        }
+        if((play_submarine_pos[1] - play_submarine_pos[0]) == 1)
+        {
+            sub_or = 0;
+        }
+        else
+        {
+            sub_or = 1;
+        }
+        if((play_cruiser_pos[1] - play_cruiser_pos[0]) == 1)
+        {
+            cru_or = 0;
+        }
+        else
+        {
+            cru_or = 1;
+        }
+        if((play_destroyer_pos[1] - play_destroyer_pos[0]) == 1)
+        {
+            des_or = 0;
+        }
+        else
+        {
+            des_or = 1;
+        }
+        System.out.println("Aircraft Carrier Orientation : "+air_or);
+        for(int i = 0 ; i < 5 ; i ++)
+        {
+            System.out.println(play_aircraft_pos[i]);
+        }
+        System.out.println("Battleship Orientation : "+bat_or);
+        for(int i = 0 ; i < 4 ; i ++)
+        {
+            System.out.println(play_battleship_pos[i]);
+        }
+        System.out.println("Submarine Orientation : "+sub_or);
+        for(int i = 0 ; i < 3 ; i ++)
+        {
+            System.out.println(play_submarine_pos[i]);
+        }
+        System.out.println("Cruiser Orientation : "+cru_or);
+        for(int i = 0 ; i < 3 ; i ++)
+        {
+            System.out.println(play_cruiser_pos[i]);
+        }
+        System.out.println("Destroyer Orientation : "+des_or);
+        for(int i = 0 ; i < 2 ; i ++)
+        {
+            System.out.println(play_destroyer_pos[i]);
+        }
+    }
+}
